Network.AWS.ECS

Description

Amazon EC2 Container Service is a highly scalable, high performance container management service that supports Docker containers and allows you to easily run distributed applications on a managed cluster of Amazon EC2 instances. Amazon EC2 Container Service lets you launch and stop container-enabled applications with simple API calls, allows you to query the state of your cluster from a centralized service, and gives you access to many familiar Amazon EC2 features like security groups, EBS volumes and IAM roles. You can use EC2 Container Service to schedule the placement of containers across your cluster based on your resource needs, isolation policies, and availability requirements. Amazon EC2 Container Service eliminates the need for you to operate your own cluster management and configuration management systems or worry about scaling your management infrastructure.

Documentation

module Network.AWS.ECS.CreateCluster

module Network.AWS.ECS.CreateService

module Network.AWS.ECS.DeleteCluster

module Network.AWS.ECS.DeleteService

module Network.AWS.ECS.DeregisterContainerInstance

module Network.AWS.ECS.DeregisterTaskDefinition

module Network.AWS.ECS.DescribeClusters

module Network.AWS.ECS.DescribeContainerInstances

module Network.AWS.ECS.DescribeServices

module Network.AWS.ECS.DescribeTaskDefinition

module Network.AWS.ECS.DescribeTasks

module Network.AWS.ECS.DiscoverPollEndpoint

module Network.AWS.ECS.ListClusters

module Network.AWS.ECS.ListContainerInstances

module Network.AWS.ECS.ListServices

module Network.AWS.ECS.ListTaskDefinitionFamilies

module Network.AWS.ECS.ListTaskDefinitions

module Network.AWS.ECS.ListTasks

module Network.AWS.ECS.RegisterContainerInstance

module Network.AWS.ECS.RegisterTaskDefinition

module Network.AWS.ECS.RunTask

module Network.AWS.ECS.StartTask

module Network.AWS.ECS.StopTask

module Network.AWS.ECS.SubmitContainerStateChange

module Network.AWS.ECS.SubmitTaskStateChange

module Network.AWS.ECS.Types

module Network.AWS.ECS.UpdateService