Network.AWS.ECS.CreateCluster

Description

Creates a new Amazon ECS cluster. By default, your account will receive a 'default' cluster when you launch your first container instance. However, you can create your own cluster with a unique name with the CreateCluster action.

http://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_CreateCluster.html

Request

data CreateCluster

Request constructor

createCluster :: CreateCluster

CreateCluster constructor.

The fields accessible through corresponding lenses are:

ccClusterName :: Maybe Text

Request lenses

ccClusterName :: Lens' CreateCluster (Maybe Text)

The name of your cluster. If you do not specify a name for your cluster, you will create a cluster named 'default'. Up to 255 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed.

Response

data CreateClusterResponse

Response constructor

createClusterResponse :: CreateClusterResponse

CreateClusterResponse constructor.

The fields accessible through corresponding lenses are:

ccrCluster :: Maybe Cluster

Response lenses

ccrCluster :: Lens' CreateClusterResponse (Maybe Cluster)

The full description of your new cluster.