Network.AWS.ECS.CreateService

Description

Runs and maintains a desired number of tasks from a specified task definition. If the number of tasks running in a service drops below desiredCount, Amazon ECS will spawn another instantiation of the task in the specified cluster.

http://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_CreateService.html

Request

data CreateService

Request constructor

createService

Arguments

:: Text
cs1ServiceName

-> CreateService

CreateService constructor.

The fields accessible through corresponding lenses are:

cs1ClientToken :: Maybe Text

cs1Cluster :: Maybe Text

cs1DesiredCount :: Maybe Int

cs1LoadBalancers :: [LoadBalancer]

cs1Role :: Maybe Text

cs1ServiceName :: Text

cs1TaskDefinition :: Maybe Text

Request lenses

cs1ClientToken :: Lens' CreateService (Maybe Text)

Unique, case-sensitive identifier you provide to ensure the idempotency of the request. Up to 32 ASCII characters are allowed.

cs1Cluster :: Lens' CreateService (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster that you want to run your service on. If you do not specify a cluster, the default cluster is assumed.

cs1DesiredCount :: Lens' CreateService (Maybe Int)

The number of instantiations of the specified task definition that you would like to place and keep running on your cluster.

cs1LoadBalancers :: Lens' CreateService [LoadBalancer]

A list of load balancer objects, containing the load balancer name, the container name (as it appears in a container definition), and the container port to access from the load balancer.

cs1Role :: Lens' CreateService (Maybe Text)

The name or full Amazon Resource Name (ARN) of the IAM role that allows your Amazon ECS container agent to make calls to your load balancer on your behalf. This parameter is only required if you are using a load balancer with your service.

cs1ServiceName :: Lens' CreateService Text

The name of your service. Up to 255 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed.

cs1TaskDefinition :: Lens' CreateService (Maybe Text)

The family and revision ('family:revision') or full Amazon Resource Name (ARN) of the task definition that you want to run in your service.

Response

data CreateServiceResponse

Response constructor

createServiceResponse :: CreateServiceResponse

CreateServiceResponse constructor.

The fields accessible through corresponding lenses are:

csrService :: Maybe ContainerService

Response lenses

csrService :: Lens' CreateServiceResponse (Maybe ContainerService)

The full description of your service following the create call.