Network.AWS.ECS.DeleteCluster

Description

Deletes the specified cluster. You must deregister all container instances from this cluster before you may delete it. You can list the container instances in a cluster with ListContainerInstances and deregister them with DeregisterContainerInstance.

http://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_DeleteCluster.html

Request

data DeleteCluster

Request constructor

deleteCluster

Arguments

:: Text
dcCluster

-> DeleteCluster

DeleteCluster constructor.

The fields accessible through corresponding lenses are:

dcCluster :: Text

Request lenses

dcCluster :: Lens' DeleteCluster Text

The short name or full Amazon Resource Name (ARN) of the cluster that you want to delete.

Response

data DeleteClusterResponse

Response constructor

deleteClusterResponse :: DeleteClusterResponse

DeleteClusterResponse constructor.

The fields accessible through corresponding lenses are:

dcrCluster :: Maybe Cluster

Response lenses

dcrCluster :: Lens' DeleteClusterResponse (Maybe Cluster)

The full description of the deleted cluster.