Network.AWS.ECS.DeleteService

Description

Deletes a specified service within a cluster.

http://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_DeleteService.html

Request

data DeleteService

Request constructor

deleteService

Arguments

:: Text
dsService

-> DeleteService

DeleteService constructor.

The fields accessible through corresponding lenses are:

dsCluster :: Maybe Text

dsService :: Text

Request lenses

dsCluster :: Lens' DeleteService (Maybe Text)

The name of the cluster that hosts the service you want to delete.

dsService :: Lens' DeleteService Text

The name of the service you want to delete.

Response

data DeleteServiceResponse

Response constructor

deleteServiceResponse :: DeleteServiceResponse

DeleteServiceResponse constructor.

The fields accessible through corresponding lenses are:

dsrService :: Maybe ContainerService

Response lenses

dsrService :: Lens' DeleteServiceResponse (Maybe ContainerService)