Network.AWS.ECS.DeregisterContainerInstance

Description

Deregisters an Amazon ECS container instance from the specified cluster. This instance will no longer be available to run tasks.

http://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_DeregisterContainerInstance.html

Request

data DeregisterContainerInstance

Request constructor

deregisterContainerInstance

Arguments

:: Text
dci1ContainerInstance

-> DeregisterContainerInstance

DeregisterContainerInstance constructor.

The fields accessible through corresponding lenses are:

dci1Cluster :: Maybe Text

dci1ContainerInstance :: Text

dci1Force :: Maybe Bool

Request lenses

dci1Cluster :: Lens' DeregisterContainerInstance (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the container instance you want to deregister. If you do not specify a cluster, the default cluster is assumed.

dci1ContainerInstance :: Lens' DeregisterContainerInstance Text

The container instance UUID or full Amazon Resource Name (ARN) of the container instance you want to deregister. The ARN contains the 'arn:aws:ecs' namespace, followed by the region of the container instance, the AWS account ID of the container instance owner, the 'container-instance' namespace, and then the container instance UUID. For example, arn:aws:ecs:region:aws_account_id:container-instance/container_instance_UUID.

dci1Force :: Lens' DeregisterContainerInstance (Maybe Bool)

Force the deregistration of the container instance. You can use the force parameter if you have several tasks running on a container instance and you don't want to run StopTask for each task before deregistering the container instance.

Response

data DeregisterContainerInstanceResponse

Response constructor

deregisterContainerInstanceResponse :: DeregisterContainerInstanceResponse

DeregisterContainerInstanceResponse constructor.

The fields accessible through corresponding lenses are:

dcirContainerInstance :: Maybe ContainerInstance

Response lenses

dcirContainerInstance :: Lens' DeregisterContainerInstanceResponse (Maybe ContainerInstance)