Network.AWS.ECS.DeregisterTaskDefinition

Description

NOT YET IMPLEMENTED.

Deregisters the specified task definition. You will no longer be able to run tasks from this definition after deregistration.

http://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_DeregisterTaskDefinition.html

Request

data DeregisterTaskDefinition

Request constructor

deregisterTaskDefinition

Arguments

:: Text
dtd1TaskDefinition

-> DeregisterTaskDefinition

DeregisterTaskDefinition constructor.

The fields accessible through corresponding lenses are:

dtd1TaskDefinition :: Text

Request lenses

dtd1TaskDefinition :: Lens' DeregisterTaskDefinition Text

The family and revision ('family:revision') or full Amazon Resource Name (ARN) of the task definition that you want to deregister.

Response

data DeregisterTaskDefinitionResponse

Response constructor

deregisterTaskDefinitionResponse :: DeregisterTaskDefinitionResponse

DeregisterTaskDefinitionResponse constructor.

The fields accessible through corresponding lenses are:

dtdrTaskDefinition :: Maybe TaskDefinition

Response lenses

dtdrTaskDefinition :: Lens' DeregisterTaskDefinitionResponse (Maybe TaskDefinition)

The full description of the deregistered task.