Network.AWS.ECS.DescribeContainerInstances

Description

Describes Amazon EC2 Container Service container instances. Returns metadata about registered and remaining resources on each container instance requested.

Creating a Request

describeContainerInstances :: DescribeContainerInstances

Creates a value of DescribeContainerInstances with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dciCluster - The short name or full Amazon Resource Name (ARN) of the cluster that hosts the container instances to describe. If you do not specify a cluster, the default cluster is assumed.

dciContainerInstances - A list of container instance IDs or full Amazon Resource Name (ARN) entries.

data DescribeContainerInstances

See: describeContainerInstances smart constructor.

Request Lenses

dciCluster :: Lens' DescribeContainerInstances (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the container instances to describe. If you do not specify a cluster, the default cluster is assumed.

dciContainerInstances :: Lens' DescribeContainerInstances [Text]

A list of container instance IDs or full Amazon Resource Name (ARN) entries.

Destructuring the Response

describeContainerInstancesResponse

Arguments

:: Int
dcisrsResponseStatus

-> DescribeContainerInstancesResponse

Creates a value of DescribeContainerInstancesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcisrsFailures - Any failures associated with the call.

dcisrsContainerInstances - The list of container instances.

dcisrsResponseStatus - -- | The response status code.

data DescribeContainerInstancesResponse

See: describeContainerInstancesResponse smart constructor.

Response Lenses

dcisrsFailures :: Lens' DescribeContainerInstancesResponse [Failure]

Any failures associated with the call.

dcisrsContainerInstances :: Lens' DescribeContainerInstancesResponse [ContainerInstance]

The list of container instances.

dcisrsResponseStatus :: Lens' DescribeContainerInstancesResponse Int

- | The response status code.