Network.AWS.ECS.DescribeTaskDefinition

Description

Describes a task definition. You can specify a family and revision to find information on a specific task definition, or you can simply specify the family to find the latest revision in that family.

http://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_DescribeTaskDefinition.html

Request

data DescribeTaskDefinition

Request constructor

describeTaskDefinition

Arguments

:: Text
dtdTaskDefinition

-> DescribeTaskDefinition

DescribeTaskDefinition constructor.

The fields accessible through corresponding lenses are:

dtdTaskDefinition :: Text

Request lenses

dtdTaskDefinition :: Lens' DescribeTaskDefinition Text

The family for the latest revision, family and revision ('family:revision') for a specific revision in the family, or full Amazon Resource Name (ARN) of the task definition that you want to describe.

Response

data DescribeTaskDefinitionResponse

Response constructor

describeTaskDefinitionResponse :: DescribeTaskDefinitionResponse

DescribeTaskDefinitionResponse constructor.

The fields accessible through corresponding lenses are:

dtdr1TaskDefinition :: Maybe TaskDefinition

Response lenses

dtdr1TaskDefinition :: Lens' DescribeTaskDefinitionResponse (Maybe TaskDefinition)

The full task definition description.