Network.AWS.ECS.DescribeTasks

Description

Describes a specified task or tasks.

http://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_DescribeTasks.html

Request

data DescribeTasks

Request constructor

describeTasks :: DescribeTasks

DescribeTasks constructor.

The fields accessible through corresponding lenses are:

dtCluster :: Maybe Text

dtTasks :: [Text]

Request lenses

dtCluster :: Lens' DescribeTasks (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the task you want to describe. If you do not specify a cluster, the default cluster is assumed.

dtTasks :: Lens' DescribeTasks [Text]

A space-separated list of task UUIDs or full Amazon Resource Name (ARN) entries.

Response

data DescribeTasksResponse

Response constructor

describeTasksResponse :: DescribeTasksResponse

DescribeTasksResponse constructor.

The fields accessible through corresponding lenses are:

dtrFailures :: [Failure]

dtrTasks :: [Task]

Response lenses

dtrFailures :: Lens' DescribeTasksResponse [Failure]

dtrTasks :: Lens' DescribeTasksResponse [Task]

The list of tasks.