Network.AWS.ECS.DiscoverPollEndpoint

Description

This action is only used by the Amazon EC2 Container Service agent, and it is not intended for use outside of the agent.

Returns an endpoint for the Amazon EC2 Container Service agent to poll for updates.

http://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_DiscoverPollEndpoint.html

Request

data DiscoverPollEndpoint

Request constructor

discoverPollEndpoint :: DiscoverPollEndpoint

DiscoverPollEndpoint constructor.

The fields accessible through corresponding lenses are:

dpeCluster :: Maybe Text

dpeContainerInstance :: Maybe Text

Request lenses

dpeCluster :: Lens' DiscoverPollEndpoint (Maybe Text)

The cluster that the container instance belongs to.

dpeContainerInstance :: Lens' DiscoverPollEndpoint (Maybe Text)

The container instance UUID or full Amazon Resource Name (ARN) of the container instance. The ARN contains the 'arn:aws:ecs' namespace, followed by the region of the container instance, the AWS account ID of the container instance owner, the 'container-instance' namespace, and then the container instance UUID. For example, arn:aws:ecs:region:aws_account_id:container-instance/ container_instance_UUID.

Response

data DiscoverPollEndpointResponse

Response constructor

discoverPollEndpointResponse :: DiscoverPollEndpointResponse

DiscoverPollEndpointResponse constructor.

The fields accessible through corresponding lenses are:

dperEndpoint :: Maybe Text

Response lenses

dperEndpoint :: Lens' DiscoverPollEndpointResponse (Maybe Text)

The endpoint for the Amazon ECS agent to poll.