Network.AWS.ECS.ListAttributes

Description

Lists the attributes for Amazon ECS resources within a specified target type and cluster. When you specify a target type and cluster, ListAttributes returns a list of attribute objects, one for each attribute on each resource. You can filter the list of results to a single attribute name to only return results that have that name. You can also filter the results by attribute name and value, for example, to see which container instances in a cluster are running a Linux AMI (ecs.os-type=linux ).

Creating a Request

listAttributes

Arguments

:: TargetType
laTargetType

-> ListAttributes

Creates a value of ListAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

laAttributeValue - The value of the attribute with which to filter results. You must also specify an attribute name to use this parameter.

laCluster - The short name or full Amazon Resource Name (ARN) of the cluster to list attributes. If you do not specify a cluster, the default cluster is assumed.

laNextToken - The nextToken value returned from a previous paginated ListAttributes request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

laAttributeName - The name of the attribute with which to filter the results.

laMaxResults - The maximum number of cluster results returned by ListAttributes in paginated output. When this parameter is used, ListAttributes only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListAttributes request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListAttributes returns up to 100 results and a nextToken value if applicable.

laTargetType - The type of the target with which to list attributes.

data ListAttributes

See: listAttributes smart constructor.

Request Lenses

laAttributeValue :: Lens' ListAttributes (Maybe Text)

The value of the attribute with which to filter results. You must also specify an attribute name to use this parameter.

laCluster :: Lens' ListAttributes (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster to list attributes. If you do not specify a cluster, the default cluster is assumed.

laNextToken :: Lens' ListAttributes (Maybe Text)

The nextToken value returned from a previous paginated ListAttributes request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

laAttributeName :: Lens' ListAttributes (Maybe Text)

The name of the attribute with which to filter the results.

laMaxResults :: Lens' ListAttributes (Maybe Int)

The maximum number of cluster results returned by ListAttributes in paginated output. When this parameter is used, ListAttributes only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListAttributes request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListAttributes returns up to 100 results and a nextToken value if applicable.

laTargetType :: Lens' ListAttributes TargetType

The type of the target with which to list attributes.

Destructuring the Response

listAttributesResponse

Arguments

:: Int
larsResponseStatus

-> ListAttributesResponse

Creates a value of ListAttributesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

larsNextToken - The nextToken value to include in a future ListAttributes request. When the results of a ListAttributes request exceed maxResults , this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

larsAttributes - A list of attribute objects that meet the criteria of the request.

larsResponseStatus - -- | The response status code.

data ListAttributesResponse

See: listAttributesResponse smart constructor.

Response Lenses

larsNextToken :: Lens' ListAttributesResponse (Maybe Text)

The nextToken value to include in a future ListAttributes request. When the results of a ListAttributes request exceed maxResults , this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

larsAttributes :: Lens' ListAttributesResponse [Attribute]

A list of attribute objects that meet the criteria of the request.

larsResponseStatus :: Lens' ListAttributesResponse Int

- | The response status code.