Network.AWS.ECS.ListClusters

Description

Returns a list of existing clusters.

http://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_ListClusters.html

Request

data ListClusters

Request constructor

listClusters :: ListClusters

ListClusters constructor.

The fields accessible through corresponding lenses are:

lcMaxResults :: Maybe Int

lcNextToken :: Maybe Text

Request lenses

lcMaxResults :: Lens' ListClusters (Maybe Int)

The maximum number of cluster results returned by ListClusters in paginated output. When this parameter is used, ListClusters only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListClusters request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListClusters returns up to 100 results and a nextToken value if applicable.

lcNextToken :: Lens' ListClusters (Maybe Text)

The nextToken value returned from a previous paginated ListClusters request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

Response

data ListClustersResponse

Response constructor

listClustersResponse :: ListClustersResponse

ListClustersResponse constructor.

The fields accessible through corresponding lenses are:

lcrClusterArns :: [Text]

lcrNextToken :: Maybe Text

Response lenses

lcrClusterArns :: Lens' ListClustersResponse [Text]

The list of full Amazon Resource Name (ARN) entries for each cluster associated with your account.

lcrNextToken :: Lens' ListClustersResponse (Maybe Text)

The nextToken value to include in a future ListClusters request. When the results of a ListClusters request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.