Network.AWS.ECS.ListContainerInstances

Description

Returns a list of container instances in a specified cluster.

http://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_ListContainerInstances.html

Request

data ListContainerInstances

Request constructor

listContainerInstances :: ListContainerInstances

ListContainerInstances constructor.

The fields accessible through corresponding lenses are:

lciCluster :: Maybe Text

lciMaxResults :: Maybe Int

lciNextToken :: Maybe Text

Request lenses

lciCluster :: Lens' ListContainerInstances (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the container instances you want to list. If you do not specify a cluster, the default cluster is assumed..

lciMaxResults :: Lens' ListContainerInstances (Maybe Int)

The maximum number of container instance results returned by ListContainerInstances in paginated output. When this parameter is used, ListContainerInstances only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListContainerInstances request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListContainerInstances returns up to 100 results and a nextToken value if applicable.

lciNextToken :: Lens' ListContainerInstances (Maybe Text)

The nextToken value returned from a previous paginated ListContainerInstances request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

Response

data ListContainerInstancesResponse

Response constructor

listContainerInstancesResponse :: ListContainerInstancesResponse

ListContainerInstancesResponse constructor.

The fields accessible through corresponding lenses are:

lcirContainerInstanceArns :: [Text]

lcirNextToken :: Maybe Text

Response lenses

lcirContainerInstanceArns :: Lens' ListContainerInstancesResponse [Text]

The list of container instance full Amazon Resource Name (ARN) entries for each container instance associated with the specified cluster.

lcirNextToken :: Lens' ListContainerInstancesResponse (Maybe Text)

The nextToken value to include in a future ListContainerInstances request. When the results of a ListContainerInstances request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.