Network.AWS.ECS.ListServices

Description

Lists the services that are running in a specified cluster.

http://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_ListServices.html

Request

data ListServices

Request constructor

listServices :: ListServices

ListServices constructor.

The fields accessible through corresponding lenses are:

lsCluster :: Maybe Text

lsMaxResults :: Maybe Int

lsNextToken :: Maybe Text

Request lenses

lsCluster :: Lens' ListServices (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the services you want to list. If you do not specify a cluster, the default cluster is assumed..

lsMaxResults :: Lens' ListServices (Maybe Int)

The maximum number of container instance results returned by ListServices in paginated output. When this parameter is used, ListServices only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListServices request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListServices returns up to 100 results and a nextToken value if applicable.

lsNextToken :: Lens' ListServices (Maybe Text)

The nextToken value returned from a previous paginated ListServices request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

Response

data ListServicesResponse

Response constructor

listServicesResponse :: ListServicesResponse

ListServicesResponse constructor.

The fields accessible through corresponding lenses are:

lsrNextToken :: Maybe Text

lsrServiceArns :: [Text]

Response lenses

lsrNextToken :: Lens' ListServicesResponse (Maybe Text)

The nextToken value to include in a future ListServices request. When the results of a ListServices request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

lsrServiceArns :: Lens' ListServicesResponse [Text]

The list of full Amazon Resource Name (ARN) entries for each service associated with the specified cluster.