Network.AWS.ECS.ListTaskDefinitionFamilies

Description

Returns a list of task definition families that are registered to your account. You can filter the results with the familyPrefix parameter.

http://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_ListTaskDefinitionFamilies.html

Request

data ListTaskDefinitionFamilies

Request constructor

listTaskDefinitionFamilies :: ListTaskDefinitionFamilies

ListTaskDefinitionFamilies constructor.

The fields accessible through corresponding lenses are:

ltdfFamilyPrefix :: Maybe Text

ltdfMaxResults :: Maybe Int

ltdfNextToken :: Maybe Text

Request lenses

ltdfFamilyPrefix :: Lens' ListTaskDefinitionFamilies (Maybe Text)

The familyPrefix is a string that is used to filter the results of ListTaskDefinitionFamilies. If you specify a familyPrefix, only task definition family names that begin with the familyPrefix string are returned.

ltdfMaxResults :: Lens' ListTaskDefinitionFamilies (Maybe Int)

The maximum number of task definition family results returned by ListTaskDefinitionFamilies in paginated output. When this parameter is used, ListTaskDefinitions only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListTaskDefinitionFamilies request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListTaskDefinitionFamilies returns up to 100 results and a nextToken value if applicable.

ltdfNextToken :: Lens' ListTaskDefinitionFamilies (Maybe Text)

The nextToken value returned from a previous paginated ListTaskDefinitionFamilies request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

Response

data ListTaskDefinitionFamiliesResponse

Response constructor

listTaskDefinitionFamiliesResponse :: ListTaskDefinitionFamiliesResponse

ListTaskDefinitionFamiliesResponse constructor.

The fields accessible through corresponding lenses are:

ltdfrFamilies :: [Text]

ltdfrNextToken :: Maybe Text

Response lenses

ltdfrFamilies :: Lens' ListTaskDefinitionFamiliesResponse [Text]

The list of task definition family names that match the ListTaskDefinitionFamilies request.

ltdfrNextToken :: Lens' ListTaskDefinitionFamiliesResponse (Maybe Text)

The nextToken value to include in a future ListTaskDefinitionFamilies request. When the results of a ListTaskDefinitionFamilies request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.