Network.AWS.ECS.ListTaskDefinitions

Description

Returns a list of task definitions that are registered to your account. You can filter the results by family name with the familyPrefix parameter.

http://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_ListTaskDefinitions.html

Request

data ListTaskDefinitions

Request constructor

listTaskDefinitions :: ListTaskDefinitions

ListTaskDefinitions constructor.

The fields accessible through corresponding lenses are:

ltdFamilyPrefix :: Maybe Text

ltdMaxResults :: Maybe Int

ltdNextToken :: Maybe Text

Request lenses

ltdFamilyPrefix :: Lens' ListTaskDefinitions (Maybe Text)

The full family name that you want to filter the ListTaskDefinitions results with. Specifying a familyPrefix will limit the listed task definitions to task definition revisions that belong to that family.

ltdMaxResults :: Lens' ListTaskDefinitions (Maybe Int)

The maximum number of task definition results returned by ListTaskDefinitions in paginated output. When this parameter is used, ListTaskDefinitions only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListTaskDefinitions request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListTaskDefinitions returns up to 100 results and a nextToken value if applicable.

ltdNextToken :: Lens' ListTaskDefinitions (Maybe Text)

The nextToken value returned from a previous paginated ListTaskDefinitions request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

Response

data ListTaskDefinitionsResponse

Response constructor

listTaskDefinitionsResponse :: ListTaskDefinitionsResponse

ListTaskDefinitionsResponse constructor.

The fields accessible through corresponding lenses are:

ltdrNextToken :: Maybe Text

ltdrTaskDefinitionArns :: [Text]

Response lenses

ltdrNextToken :: Lens' ListTaskDefinitionsResponse (Maybe Text)

The nextToken value to include in a future ListTaskDefinitions request. When the results of a ListTaskDefinitions request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

ltdrTaskDefinitionArns :: Lens' ListTaskDefinitionsResponse [Text]

The list of task definition Amazon Resource Name (ARN) entries for the ListTaskDefintions request.