Network.AWS.ECS.ListTasks

Description

Returns a list of tasks for a specified cluster. You can filter the results by family name or by a particular container instance with the family and containerInstance parameters.

http://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_ListTasks.html

Request

data ListTasks

Request constructor

listTasks :: ListTasks

ListTasks constructor.

The fields accessible through corresponding lenses are:

ltCluster :: Maybe Text

ltContainerInstance :: Maybe Text

ltFamily :: Maybe Text

ltMaxResults :: Maybe Int

ltNextToken :: Maybe Text

ltServiceName :: Maybe Text

ltStartedBy :: Maybe Text

Request lenses

ltCluster :: Lens' ListTasks (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the tasks you want to list. If you do not specify a cluster, the default cluster is assumed..

ltContainerInstance :: Lens' ListTasks (Maybe Text)

The container instance UUID or full Amazon Resource Name (ARN) of the container instance that you want to filter the ListTasks results with. Specifying a containerInstance will limit the results to tasks that belong to that container instance.

ltFamily :: Lens' ListTasks (Maybe Text)

The name of the family that you want to filter the ListTasks results with. Specifying a family will limit the results to tasks that belong to that family.

ltMaxResults :: Lens' ListTasks (Maybe Int)

The maximum number of task results returned by ListTasks in paginated output. When this parameter is used, ListTasks only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListTasks request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListTasks returns up to 100 results and a nextToken value if applicable.

ltNextToken :: Lens' ListTasks (Maybe Text)

The nextToken value returned from a previous paginated ListTasks request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

ltServiceName :: Lens' ListTasks (Maybe Text)

The name of the service that you want to filter the ListTasks results with. Specifying a serviceName will limit the results to tasks that belong to that service.

ltStartedBy :: Lens' ListTasks (Maybe Text)

The startedBy value that you want to filter the task results with. Specifying a startedBy value will limit the results to tasks that were started with that value.

Response

data ListTasksResponse

Response constructor

listTasksResponse :: ListTasksResponse

ListTasksResponse constructor.

The fields accessible through corresponding lenses are:

ltrNextToken :: Maybe Text

ltrTaskArns :: [Text]

Response lenses

ltrNextToken :: Lens' ListTasksResponse (Maybe Text)

The nextToken value to include in a future ListTasks request. When the results of a ListTasks request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

ltrTaskArns :: Lens' ListTasksResponse [Text]

The list of task Amazon Resource Name (ARN) entries for the ListTasks request.