Network.AWS.ECS.PutAttributes

Description

Create or update an attribute on an Amazon ECS resource. If the attribute does not exist, it is created. If the attribute exists, its value is replaced with the specified value. To delete an attribute, use DeleteAttributes . For more information, see Attributes in the Amazon EC2 Container Service Developer Guide .

Creating a Request

putAttributes :: PutAttributes

Creates a value of PutAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

paCluster - The short name or full Amazon Resource Name (ARN) of the cluster that contains the resource to apply attributes. If you do not specify a cluster, the default cluster is assumed.

paAttributes - The attributes to apply to your resource. You can specify up to 10 custom attributes per resource. You can specify up to 10 attributes in a single call.

data PutAttributes

See: putAttributes smart constructor.

Request Lenses

paCluster :: Lens' PutAttributes (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster that contains the resource to apply attributes. If you do not specify a cluster, the default cluster is assumed.

paAttributes :: Lens' PutAttributes [Attribute]

The attributes to apply to your resource. You can specify up to 10 custom attributes per resource. You can specify up to 10 attributes in a single call.

Destructuring the Response

putAttributesResponse

Arguments

:: Int
parsResponseStatus

-> PutAttributesResponse

Creates a value of PutAttributesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

parsAttributes - The attributes applied to your resource.

parsResponseStatus - -- | The response status code.

data PutAttributesResponse

See: putAttributesResponse smart constructor.

Response Lenses

parsAttributes :: Lens' PutAttributesResponse [Attribute]

The attributes applied to your resource.

parsResponseStatus :: Lens' PutAttributesResponse Int

- | The response status code.