Network.AWS.ECS.RegisterContainerInstance

Description

This action is only used by the Amazon EC2 Container Service agent, and it is not intended for use outside of the agent.

Registers an Amazon EC2 instance into the specified cluster. This instance will become available to place containers on.

http://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_RegisterContainerInstance.html

Request

data RegisterContainerInstance

Request constructor

registerContainerInstance :: RegisterContainerInstance

RegisterContainerInstance constructor.

The fields accessible through corresponding lenses are:

rciCluster :: Maybe Text

rciInstanceIdentityDocument :: Maybe Text

rciInstanceIdentityDocumentSignature :: Maybe Text

rciTotalResources :: [Resource]

rciVersionInfo :: Maybe VersionInfo

Request lenses

rciCluster :: Lens' RegisterContainerInstance (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster that you want to register your container instance with. If you do not specify a cluster, the default cluster is assumed..

rciInstanceIdentityDocument :: Lens' RegisterContainerInstance (Maybe Text)

rciInstanceIdentityDocumentSignature :: Lens' RegisterContainerInstance (Maybe Text)

rciTotalResources :: Lens' RegisterContainerInstance [Resource]

rciVersionInfo :: Lens' RegisterContainerInstance (Maybe VersionInfo)

Response

data RegisterContainerInstanceResponse

Response constructor

registerContainerInstanceResponse :: RegisterContainerInstanceResponse

RegisterContainerInstanceResponse constructor.

The fields accessible through corresponding lenses are:

rcirContainerInstance :: Maybe ContainerInstance

Response lenses

rcirContainerInstance :: Lens' RegisterContainerInstanceResponse (Maybe ContainerInstance)