Network.AWS.ECS.RegisterTaskDefinition

Description

Registers a new task definition from the supplied family and containerDefinitions. Optionally, you can add data volumes to your containers with the volumes parameter. For more information on task definition parameters and defaults, see Amazon ECS Task Definitions in the Amazon EC2 Container Service DeveloperGuide.

http://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_RegisterTaskDefinition.html

Request

data RegisterTaskDefinition

Request constructor

registerTaskDefinition

Arguments

:: Text
rtdFamily

-> RegisterTaskDefinition

RegisterTaskDefinition constructor.

The fields accessible through corresponding lenses are:

rtdContainerDefinitions :: [ContainerDefinition]

rtdFamily :: Text

rtdVolumes :: [Volume]

Request lenses

rtdContainerDefinitions :: Lens' RegisterTaskDefinition [ContainerDefinition]

A list of container definitions in JSON format that describe the different containers that make up your task.

rtdFamily :: Lens' RegisterTaskDefinition Text

You must specify a family for a task definition, which allows you to track multiple versions of the same task definition. You can think of the family as a name for your task definition. Up to 255 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed.

rtdVolumes :: Lens' RegisterTaskDefinition [Volume]

A list of volume definitions in JSON format that containers in your task may use.

Response

data RegisterTaskDefinitionResponse

Response constructor

registerTaskDefinitionResponse :: RegisterTaskDefinitionResponse

RegisterTaskDefinitionResponse constructor.

The fields accessible through corresponding lenses are:

rtdrTaskDefinition :: Maybe TaskDefinition

Response lenses

rtdrTaskDefinition :: Lens' RegisterTaskDefinitionResponse (Maybe TaskDefinition)