Network.AWS.ECS.RunTask

Description

Start a task using random placement and the default Amazon ECS scheduler. If you want to use your own scheduler or place a task on a specific container instance, use StartTask instead.

The count parameter is limited to 10 tasks per call.

http://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_RunTask.html

Request

data RunTask

Request constructor

runTask

Arguments

:: Text
rtTaskDefinition

-> RunTask

RunTask constructor.

The fields accessible through corresponding lenses are:

rtCluster :: Maybe Text

rtCount :: Maybe Int

rtOverrides :: Maybe TaskOverride

rtStartedBy :: Maybe Text

rtTaskDefinition :: Text

Request lenses

rtCluster :: Lens' RunTask (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster that you want to run your task on. If you do not specify a cluster, the default cluster is assumed..

rtCount :: Lens' RunTask (Maybe Int)

The number of instantiations of the specified task that you would like to place on your cluster.

The count parameter is limited to 10 tasks per call.

rtOverrides :: Lens' RunTask (Maybe TaskOverride)

A list of container overrides in JSON format that specify the name of a container in the specified task definition and the command it should run instead of its default. A total of 8192 characters are allowed for overrides. This limit includes the JSON formatting characters of the override structure.

rtStartedBy :: Lens' RunTask (Maybe Text)

An optional tag specified when a task is started. For example if you automatically trigger a task to run a batch process job, you could apply a unique identifier for that job to your task with the startedBy parameter. You can then identify which tasks belong to that job by filtering the results of a ListTasks call with the startedBy value.

If a task is started by an Amazon ECS service, then the startedBy parameter contains the deployment ID of the service that starts it.

rtTaskDefinition :: Lens' RunTask Text

The family and revision ('family:revision') or full Amazon Resource Name (ARN) of the task definition that you want to run.

Response

data RunTaskResponse

Response constructor

runTaskResponse :: RunTaskResponse

RunTaskResponse constructor.

The fields accessible through corresponding lenses are:

rtrFailures :: [Failure]

rtrTasks :: [Task]

Response lenses

rtrFailures :: Lens' RunTaskResponse [Failure]

Any failed tasks from your RunTask action are listed here.

rtrTasks :: Lens' RunTaskResponse [Task]

A full description of the tasks that were run. Each task that was successfully placed on your cluster will be described here.