Network.AWS.ECS.StartTask

Description

Starts a new task from the specified task definition on the specified container instance or instances.

Alternatively, you can use RunTask to place tasks for you. For more information, see Scheduling Tasks in the Amazon EC2 Container Service Developer Guide .

Creating a Request

startTask

Arguments

:: Text
sTaskDefinition

-> StartTask

Creates a value of StartTask with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sOverrides - A list of container overrides in JSON format that specify the name of a container in the specified task definition and the overrides it should receive. You can override the default command for a container (that is specified in the task definition or Docker image) with a command override. You can also override existing environment variables (that are specified in the task definition or Docker image) on a container or add new environment variables to it with an environment override.

sGroup - The name of the task group to associate with the task. The default value is the family name of the task definition (for example, family:my-family-name).

sCluster - The short name or full Amazon Resource Name (ARN) of the cluster on which to start your task. If you do not specify a cluster, the default cluster is assumed.

sStartedBy - An optional tag specified when a task is started. For example if you automatically trigger a task to run a batch process job, you could apply a unique identifier for that job to your task with the startedBy parameter. You can then identify which tasks belong to that job by filtering the results of a ListTasks call with the startedBy value. Up to 36 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed. If a task is started by an Amazon ECS service, then the startedBy parameter contains the deployment ID of the service that starts it.

sTaskDefinition - The family and revision (family:revision ) or full Amazon Resource Name (ARN) of the task definition to start. If a revision is not specified, the latest ACTIVE revision is used.

sContainerInstances - The container instance IDs or full Amazon Resource Name (ARN) entries for the container instances on which you would like to place your task. You can specify up to 10 container instances.

data StartTask

See: startTask smart constructor.

Request Lenses

sOverrides :: Lens' StartTask (Maybe TaskOverride)

A list of container overrides in JSON format that specify the name of a container in the specified task definition and the overrides it should receive. You can override the default command for a container (that is specified in the task definition or Docker image) with a command override. You can also override existing environment variables (that are specified in the task definition or Docker image) on a container or add new environment variables to it with an environment override.

sGroup :: Lens' StartTask (Maybe Text)

The name of the task group to associate with the task. The default value is the family name of the task definition (for example, family:my-family-name).

sCluster :: Lens' StartTask (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster on which to start your task. If you do not specify a cluster, the default cluster is assumed.

sStartedBy :: Lens' StartTask (Maybe Text)

An optional tag specified when a task is started. For example if you automatically trigger a task to run a batch process job, you could apply a unique identifier for that job to your task with the startedBy parameter. You can then identify which tasks belong to that job by filtering the results of a ListTasks call with the startedBy value. Up to 36 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed. If a task is started by an Amazon ECS service, then the startedBy parameter contains the deployment ID of the service that starts it.

sTaskDefinition :: Lens' StartTask Text

The family and revision (family:revision ) or full Amazon Resource Name (ARN) of the task definition to start. If a revision is not specified, the latest ACTIVE revision is used.

sContainerInstances :: Lens' StartTask [Text]

The container instance IDs or full Amazon Resource Name (ARN) entries for the container instances on which you would like to place your task. You can specify up to 10 container instances.

Destructuring the Response

startTaskResponse

Arguments

:: Int
strsResponseStatus

-> StartTaskResponse

Creates a value of StartTaskResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

strsFailures - Any failures associated with the call.

strsTasks - A full description of the tasks that were started. Each task that was successfully placed on your container instances are described here.

strsResponseStatus - -- | The response status code.

data StartTaskResponse

See: startTaskResponse smart constructor.

Response Lenses

strsFailures :: Lens' StartTaskResponse [Failure]

Any failures associated with the call.

strsTasks :: Lens' StartTaskResponse [Task]

A full description of the tasks that were started. Each task that was successfully placed on your container instances are described here.

strsResponseStatus :: Lens' StartTaskResponse Int

- | The response status code.