Network.AWS.ECS.StopTask

Description

Stops a running task.

When StopTask is called on a task, the equivalent of docker stop is issued to the containers running in the task. This results in a SIGTERM and a default 30-second timeout, after which SIGKILL is sent and the containers are forcibly stopped. If the container handles the SIGTERM gracefully and exits within 30 seconds from receiving it, no SIGKILL is sent.

Creating a Request

stopTask

Arguments

:: Text
stTask

-> StopTask

Creates a value of StopTask with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

stCluster - The short name or full Amazon Resource Name (ARN) of the cluster that hosts the task to stop. If you do not specify a cluster, the default cluster is assumed.

stReason - An optional message specified when a task is stopped. For example, if you are using a custom scheduler, you can use this parameter to specify the reason for stopping the task here, and the message will appear in subsequent DescribeTasks API operations on this task. Up to 255 characters are allowed in this message.

stTask - The task ID or full Amazon Resource Name (ARN) entry of the task to stop.

data StopTask

See: stopTask smart constructor.

Request Lenses

stCluster :: Lens' StopTask (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the task to stop. If you do not specify a cluster, the default cluster is assumed.

stReason :: Lens' StopTask (Maybe Text)

An optional message specified when a task is stopped. For example, if you are using a custom scheduler, you can use this parameter to specify the reason for stopping the task here, and the message will appear in subsequent DescribeTasks API operations on this task. Up to 255 characters are allowed in this message.

stTask :: Lens' StopTask Text

The task ID or full Amazon Resource Name (ARN) entry of the task to stop.

Destructuring the Response

stopTaskResponse

Arguments

:: Int
srsResponseStatus

-> StopTaskResponse

Creates a value of StopTaskResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

srsTask - The task that was stopped.

srsResponseStatus - -- | The response status code.

data StopTaskResponse

See: stopTaskResponse smart constructor.

Response Lenses

srsTask :: Lens' StopTaskResponse (Maybe Task)

The task that was stopped.

srsResponseStatus :: Lens' StopTaskResponse Int

- | The response status code.