Network.AWS.ECS.SubmitContainerStateChange

Description

This action is only used by the Amazon EC2 Container Service agent, and it is not intended for use outside of the agent.

Sent to acknowledge that a container changed states.

http://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_SubmitContainerStateChange.html

Request

data SubmitContainerStateChange

Request constructor

submitContainerStateChange :: SubmitContainerStateChange

SubmitContainerStateChange constructor.

The fields accessible through corresponding lenses are:

scscCluster :: Maybe Text

scscContainerName :: Maybe Text

scscExitCode :: Maybe Int

scscNetworkBindings :: [NetworkBinding]

scscReason :: Maybe Text

scscStatus :: Maybe Text

scscTask :: Maybe Text

Request lenses

scscCluster :: Lens' SubmitContainerStateChange (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the container.

scscContainerName :: Lens' SubmitContainerStateChange (Maybe Text)

The name of the container.

scscExitCode :: Lens' SubmitContainerStateChange (Maybe Int)

The exit code returned for the state change request.

scscNetworkBindings :: Lens' SubmitContainerStateChange [NetworkBinding]

The network bindings of the container.

scscReason :: Lens' SubmitContainerStateChange (Maybe Text)

The reason for the state change request.

scscStatus :: Lens' SubmitContainerStateChange (Maybe Text)

The status of the state change request.

scscTask :: Lens' SubmitContainerStateChange (Maybe Text)

The task UUID or full Amazon Resource Name (ARN) of the task that hosts the container.

Response

data SubmitContainerStateChangeResponse

Response constructor

submitContainerStateChangeResponse :: SubmitContainerStateChangeResponse

SubmitContainerStateChangeResponse constructor.

The fields accessible through corresponding lenses are:

scscrAcknowledgment :: Maybe Text

Response lenses

scscrAcknowledgment :: Lens' SubmitContainerStateChangeResponse (Maybe Text)

Acknowledgement of the state change.