Network.AWS.ECS.SubmitTaskStateChange

Description

This action is only used by the Amazon EC2 Container Service agent, and it is not intended for use outside of the agent.

Sent to acknowledge that a task changed states.

http://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_SubmitTaskStateChange.html

Request

data SubmitTaskStateChange

Request constructor

submitTaskStateChange :: SubmitTaskStateChange

SubmitTaskStateChange constructor.

The fields accessible through corresponding lenses are:

stscCluster :: Maybe Text

stscReason :: Maybe Text

stscStatus :: Maybe Text

stscTask :: Maybe Text

Request lenses

stscCluster :: Lens' SubmitTaskStateChange (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the task.

stscReason :: Lens' SubmitTaskStateChange (Maybe Text)

The reason for the state change request.

stscStatus :: Lens' SubmitTaskStateChange (Maybe Text)

The status of the state change request.

stscTask :: Lens' SubmitTaskStateChange (Maybe Text)

The task UUID or full Amazon Resource Name (ARN) of the task in the state change request.

Response

data SubmitTaskStateChangeResponse

Response constructor

submitTaskStateChangeResponse :: SubmitTaskStateChangeResponse

SubmitTaskStateChangeResponse constructor.

The fields accessible through corresponding lenses are:

stscrAcknowledgment :: Maybe Text

Response lenses

stscrAcknowledgment :: Lens' SubmitTaskStateChangeResponse (Maybe Text)

Acknowledgement of the state change.