Network.AWS.ECS.Types

Service

data ECS

Version 2014-11-13 of the Amazon EC2 Container Service service.

Error

data JSONError :: *

NetworkBinding

data NetworkBinding

networkBinding :: NetworkBinding

NetworkBinding constructor.

The fields accessible through corresponding lenses are:

nbBindIP :: Maybe Text

nbContainerPort :: Maybe Int

nbHostPort :: Maybe Int

nbBindIP :: Lens' NetworkBinding (Maybe Text)

The IP address that the container is bound to on the container instance.

nbContainerPort :: Lens' NetworkBinding (Maybe Int)

The port number on the container that is be used with the network binding.

nbHostPort :: Lens' NetworkBinding (Maybe Int)

The port number on the host that is used with the network binding.

ContainerService

data ContainerService

containerService :: ContainerService

ContainerService constructor.

The fields accessible through corresponding lenses are:

csClusterArn :: Maybe Text

csDeployments :: [Deployment]

csDesiredCount :: Maybe Int

csEvents :: [ServiceEvent]

csLoadBalancers :: [LoadBalancer]

csPendingCount :: Maybe Int

csRoleArn :: Maybe Text

csRunningCount :: Maybe Int

csServiceArn :: Maybe Text

csServiceName :: Maybe Text

csStatus :: Maybe Text

csTaskDefinition :: Maybe Text

csClusterArn :: Lens' ContainerService (Maybe Text)

The Amazon Resource Name (ARN) of the of the cluster that hosts the service.

csDeployments :: Lens' ContainerService [Deployment]

The current state of deployments for the service.

csDesiredCount :: Lens' ContainerService (Maybe Int)

The desired number of instantiations of the task definition to keep running on the service. This value is specified when the service is created with CreateService, and it can be modified with UpdateService.

csEvents :: Lens' ContainerService [ServiceEvent]

The event stream for your service. A maximum of 100 of the latest events are displayed.

csLoadBalancers :: Lens' ContainerService [LoadBalancer]

A list of load balancer objects, containing the load balancer name, the container name (as it appears in a container definition), and the container port to access from the load balancer.

csPendingCount :: Lens' ContainerService (Maybe Int)

The number of tasks in the cluster that are in the PENDING state.

csRoleArn :: Lens' ContainerService (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role associated with the service that allows the Amazon ECS container agent to register container instances with a load balancer.

csRunningCount :: Lens' ContainerService (Maybe Int)

The number of tasks in the cluster that are in the RUNNING state.

csServiceArn :: Lens' ContainerService (Maybe Text)

The Amazon Resource Name (ARN) that identifies the service. The ARN contains the 'arn:aws:ecs' namespace, followed by the region of the service, the AWS account ID of the service owner, the service namespace, and then the service name. For example, arn:aws:ecs:region:012345678910:service/my-service.

csServiceName :: Lens' ContainerService (Maybe Text)

A user-generated string that you can use to identify your service.

csStatus :: Lens' ContainerService (Maybe Text)

The status of the service. The valid values are ACTIVE, DRAINING, or INACTIVE.

csTaskDefinition :: Lens' ContainerService (Maybe Text)

The task definition to use for tasks in the service. This value is specified when the service is created with CreateService, and it can be modified with UpdateService.

Cluster

data Cluster

cluster :: Cluster

Cluster constructor.

The fields accessible through corresponding lenses are:

cClusterArn :: Maybe Text

cClusterName :: Maybe Text

cPendingTasksCount :: Maybe Int

cRegisteredContainerInstancesCount :: Maybe Int

cRunningTasksCount :: Maybe Int

cStatus :: Maybe Text

cClusterArn :: Lens' Cluster (Maybe Text)

The Amazon Resource Name (ARN) that identifies the cluster. The ARN contains the 'arn:aws:ecs' namespace, followed by the region of the cluster, the AWS account ID of the cluster owner, the cluster namespace, and then the cluster name. For example, arn:aws:ecs:region:012345678910:cluster/test.

cClusterName :: Lens' Cluster (Maybe Text)

A user-generated string that you can use to identify your cluster.

cPendingTasksCount :: Lens' Cluster (Maybe Int)

The number of tasks in the cluster that are in the PENDING state.

cRegisteredContainerInstancesCount :: Lens' Cluster (Maybe Int)

The number of container instances registered into the cluster.

cRunningTasksCount :: Lens' Cluster (Maybe Int)

The number of tasks in the cluster that are in the RUNNING state.

cStatus :: Lens' Cluster (Maybe Text)

The status of the cluster. The valid values are ACTIVE or INACTIVE. ACTIVE indicates that you can register container instances with the cluster and the associated instances can accept tasks.

Volume

data Volume

volume :: Volume

Volume constructor.

The fields accessible through corresponding lenses are:

vHost :: Maybe HostVolumeProperties

vName :: Maybe Text

vHost :: Lens' Volume (Maybe HostVolumeProperties)

The path on the host container instance that is presented to the containers which access the volume. If this parameter is empty, then the Docker daemon assigns a host path for you.

vName :: Lens' Volume (Maybe Text)

The name of the volume. This name is referenced in the sourceVolume parameter of container definition mountPoints.

ContainerOverride

data ContainerOverride

containerOverride :: ContainerOverride

ContainerOverride constructor.

The fields accessible through corresponding lenses are:

coCommand :: [Text]

coName :: Maybe Text

coCommand :: Lens' ContainerOverride [Text]

The command to send to the container that overrides the default command from the Docker image or the task definition.

coName :: Lens' ContainerOverride (Maybe Text)

The name of the container that receives the override.

KeyValuePair

data KeyValuePair

keyValuePair :: KeyValuePair

KeyValuePair constructor.

The fields accessible through corresponding lenses are:

kvpName :: Maybe Text

kvpValue :: Maybe Text

kvpName :: Lens' KeyValuePair (Maybe Text)

The name of the key value pair.

kvpValue :: Lens' KeyValuePair (Maybe Text)

The value of the key value pair.

VolumeFrom

data VolumeFrom

volumeFrom :: VolumeFrom

VolumeFrom constructor.

The fields accessible through corresponding lenses are:

vfReadOnly :: Maybe Bool

vfSourceContainer :: Maybe Text

vfReadOnly :: Lens' VolumeFrom (Maybe Bool)

If this value is true, the container has read-only access to the volume. If this value is false, then the container can write to the volume. The default value is false.

vfSourceContainer :: Lens' VolumeFrom (Maybe Text)

The name of the container to mount volumes from.

TaskOverride

data TaskOverride

taskOverride :: TaskOverride

TaskOverride constructor.

The fields accessible through corresponding lenses are:

toContainerOverrides :: [ContainerOverride]

toContainerOverrides :: Lens' TaskOverride [ContainerOverride]

One or more container overrides sent to a task.

HostVolumeProperties

data HostVolumeProperties

hostVolumeProperties :: HostVolumeProperties

HostVolumeProperties constructor.

The fields accessible through corresponding lenses are:

hvpSourcePath :: Maybe Text

hvpSourcePath :: Lens' HostVolumeProperties (Maybe Text)

The path on the host container instance that is presented to the container. If this parameter is empty, then the Docker daemon has assigned a host path for you.

VersionInfo

data VersionInfo

versionInfo :: VersionInfo

VersionInfo constructor.

The fields accessible through corresponding lenses are:

viAgentHash :: Maybe Text

viAgentVersion :: Maybe Text

viDockerVersion :: Maybe Text

viAgentHash :: Lens' VersionInfo (Maybe Text)

The Git commit hash for the Amazon ECS container agent build on the amazon-ecs-agent GitHub repository.

viAgentVersion :: Lens' VersionInfo (Maybe Text)

The version number of the Amazon ECS container agent.

viDockerVersion :: Lens' VersionInfo (Maybe Text)

The Docker version running on the container instance.

Container

data Container

container :: Container

Container constructor.

The fields accessible through corresponding lenses are:

cContainerArn :: Maybe Text

cExitCode :: Maybe Int

cLastStatus :: Maybe Text

cName :: Maybe Text

cNetworkBindings :: [NetworkBinding]

cReason :: Maybe Text

cTaskArn :: Maybe Text

cContainerArn :: Lens' Container (Maybe Text)

The Amazon Resource Name (ARN) of the container.

cExitCode :: Lens' Container (Maybe Int)

The exit code returned from the container.

cLastStatus :: Lens' Container (Maybe Text)

The last known status of the container.

cName :: Lens' Container (Maybe Text)

The name of the container.

cNetworkBindings :: Lens' Container [NetworkBinding]

cReason :: Lens' Container (Maybe Text)

A short (255 max characters) human-readable string to provide additional detail about a running or stopped container.

cTaskArn :: Lens' Container (Maybe Text)

The Amazon Resource Name (ARN) of the task.

LoadBalancer

data LoadBalancer

loadBalancer :: LoadBalancer

LoadBalancer constructor.

The fields accessible through corresponding lenses are:

lbContainerName :: Maybe Text

lbContainerPort :: Maybe Int

lbLoadBalancerName :: Maybe Text

lbContainerName :: Lens' LoadBalancer (Maybe Text)

The name of the container to associate with the load balancer.

lbContainerPort :: Lens' LoadBalancer (Maybe Int)

The port on the container to associate with the load balancer.

lbLoadBalancerName :: Lens' LoadBalancer (Maybe Text)

The name of the load balancer.

ContainerDefinition

data ContainerDefinition

containerDefinition :: ContainerDefinition

ContainerDefinition constructor.

The fields accessible through corresponding lenses are:

cdCommand :: [Text]

cdCpu :: Maybe Int

cdEntryPoint :: [Text]

cdEnvironment :: [KeyValuePair]

cdEssential :: Maybe Bool

cdImage :: Maybe Text

cdLinks :: [Text]

cdMemory :: Maybe Int

cdMountPoints :: [MountPoint]

cdName :: Maybe Text

cdPortMappings :: [PortMapping]

cdVolumesFrom :: [VolumeFrom]

cdCommand :: Lens' ContainerDefinition [Text]

The CMD that is passed to the container. For more information on the Docker CMD parameter, see https://docs.docker.com/reference/builder/#cmd.

cdCpu :: Lens' ContainerDefinition (Maybe Int)

The number of cpu units reserved for the container. A container instance has 1,024 cpu units for every CPU core.

cdEntryPoint :: Lens' ContainerDefinition [Text]

Early versions of the Amazon ECS container agent do not properly handle entryPoint parameters. If you have problems using entryPoint, update your container agent or enter your commands and arguments as command array items instead.

The ENTRYPOINT that is passed to the container. For more information on the Docker ENTRYPOINT parameter, see https://docs.docker.com/reference/builder/#entrypoint.

cdEnvironment :: Lens' ContainerDefinition [KeyValuePair]

The environment variables to pass to a container.

cdEssential :: Lens' ContainerDefinition (Maybe Bool)

If the essential parameter of a container is marked as true, the failure of that container will stop the task. If the essential parameter of a container is marked as false, then its failure will not affect the rest of the containers in a task.

cdImage :: Lens' ContainerDefinition (Maybe Text)

The image used to start a container. This string is passed directly to the Docker daemon. Images in the Docker Hub registry are available by default. Other repositories are specified with 'repository-url/image:tag.

cdLinks :: Lens' ContainerDefinition [Text]

The link parameter allows containers to communicate with each other without the need for port mappings, using the name parameter. For more information on linking Docker containers, see https://docs.docker.com/userguide/dockerlinks/.

cdMemory :: Lens' ContainerDefinition (Maybe Int)

The number of MiB of memory reserved for the container. Docker will allocate a minimum of 4 MiB of memory to a container.

cdMountPoints :: Lens' ContainerDefinition [MountPoint]

The mount points for data volumes in your container.

cdName :: Lens' ContainerDefinition (Maybe Text)

The name of a container. If you are linking multiple containers together in a task definition, the name of one container can be entered in the links of another container to connect the containers.

cdPortMappings :: Lens' ContainerDefinition [PortMapping]

The list of port mappings for the container.

cdVolumesFrom :: Lens' ContainerDefinition [VolumeFrom]

Data volumes to mount from another container.

Resource

data Resource

resource :: Resource

Resource constructor.

The fields accessible through corresponding lenses are:

rDoubleValue :: Maybe Double

rIntegerValue :: Maybe Int

rLongValue :: Maybe Integer

rName :: Maybe Text

rStringSetValue :: [Text]

rType :: Maybe Text

rDoubleValue :: Lens' Resource (Maybe Double)

When the doubleValue type is set, the value of the resource must be a double precision floating-point type.

rIntegerValue :: Lens' Resource (Maybe Int)

When the integerValue type is set, the value of the resource must be an integer.

rLongValue :: Lens' Resource (Maybe Integer)

When the longValue type is set, the value of the resource must be an extended precision floating-point type.

rName :: Lens' Resource (Maybe Text)

The name of the resource, such as CPU, MEMORY, PORTS, or a user-defined resource.

rStringSetValue :: Lens' Resource [Text]

When the stringSetValue type is set, the value of the resource must be a string type.

rType :: Lens' Resource (Maybe Text)

The type of the resource, such as INTEGER, DOUBLE, LONG, or STRINGSET.

Task

data Task

task :: Task

Task constructor.

The fields accessible through corresponding lenses are:

tClusterArn :: Maybe Text

tContainerInstanceArn :: Maybe Text

tContainers :: [Container]

tDesiredStatus :: Maybe Text

tLastStatus :: Maybe Text

tOverrides :: Maybe TaskOverride

tStartedBy :: Maybe Text

tTaskArn :: Maybe Text

tTaskDefinitionArn :: Maybe Text

tClusterArn :: Lens' Task (Maybe Text)

The Amazon Resource Name (ARN) of the of the cluster that hosts the task.

tContainerInstanceArn :: Lens' Task (Maybe Text)

The Amazon Resource Name (ARN) of the container instances that host the task.

tContainers :: Lens' Task [Container]

The containers associated with the task.

tDesiredStatus :: Lens' Task (Maybe Text)

The desired status of the task.

tLastStatus :: Lens' Task (Maybe Text)

The last known status of the task.

tOverrides :: Lens' Task (Maybe TaskOverride)

One or more container overrides.

tStartedBy :: Lens' Task (Maybe Text)

The tag specified when a task is started. If the task is started by an Amazon ECS service, then the startedBy parameter contains the deployment ID of the service that starts it.

tTaskArn :: Lens' Task (Maybe Text)

The Amazon Resource Name (ARN) of the task.

tTaskDefinitionArn :: Lens' Task (Maybe Text)

The Amazon Resource Name (ARN) of the of the task definition that creates the task.

PortMapping

data PortMapping

portMapping :: PortMapping

PortMapping constructor.

The fields accessible through corresponding lenses are:

pmContainerPort :: Maybe Int

pmHostPort :: Maybe Int

pmContainerPort :: Lens' PortMapping (Maybe Int)

The port number on the container that is bound to the user-specified or automatically assigned host port. If you specify a container port and not a host port, your container will automatically receive a host port in the 49153 to 65535 port range.

pmHostPort :: Lens' PortMapping (Maybe Int)

The port number on the container instance to reserve for your container. You can specify a non-reserved host port for your container port mapping, or you can omit the hostPort while specifying a containerPort and your container will automatically receive a port in the 49153 to 65535 port range. You should not attempt to specify a host port in the 49153 to 65535 port range, since these are reserved for automatic assignment.

The default reserved ports are 22 for SSH, the Docker ports 2375 and 2376, and the Amazon ECS Container Agent port 51678. Any host port that was previously specified in a running task is also reserved while the task is running (once a task stops, the host port is released).The current reserved ports are displayed in the remainingResources of DescribeContainerInstances output, and a container instance may have up to 50 reserved ports at a time, including the default reserved ports (automatically assigned ports do not count toward this limit).

TaskDefinition

data TaskDefinition

taskDefinition :: TaskDefinition

TaskDefinition constructor.

The fields accessible through corresponding lenses are:

tdContainerDefinitions :: [ContainerDefinition]

tdFamily :: Maybe Text

tdRevision :: Maybe Int

tdTaskDefinitionArn :: Maybe Text

tdVolumes :: [Volume]

tdContainerDefinitions :: Lens' TaskDefinition [ContainerDefinition]

A list of container definitions in JSON format that describe the different containers that make up your task. For more information on container definition parameters and defaults, see Amazon ECS Task Definitions in the Amazon EC2 Container Service Developer Guide.

tdFamily :: Lens' TaskDefinition (Maybe Text)

The family of your task definition. You can think of the family as the name of your task definition.

tdRevision :: Lens' TaskDefinition (Maybe Int)

The revision of the task in a particular family. You can think of the revision as a version number of a task definition in a family. When you register a task definition for the first time, the revision is '1', and each time you register a task definition in the same family, the revision value increases by one.

tdTaskDefinitionArn :: Lens' TaskDefinition (Maybe Text)

The full Amazon Resource Name (ARN) of the of the task definition.

tdVolumes :: Lens' TaskDefinition [Volume]

The list of volumes in a task. For more information on volume definition parameters and defaults, see Amazon ECS Task Definitions in the Amazon EC2Container Service Developer Guide.

Failure

data Failure

failure :: Failure

Failure constructor.

The fields accessible through corresponding lenses are:

fArn :: Maybe Text

fReason :: Maybe Text

fArn :: Lens' Failure (Maybe Text)

The Amazon Resource Name (ARN) of the failed resource.

fReason :: Lens' Failure (Maybe Text)

The reason for the failure.

ContainerInstance

data ContainerInstance

containerInstance :: ContainerInstance

ContainerInstance constructor.

The fields accessible through corresponding lenses are:

ciAgentConnected :: Maybe Bool

ciContainerInstanceArn :: Maybe Text

ciEc2InstanceId :: Maybe Text

ciPendingTasksCount :: Maybe Int

ciRegisteredResources :: [Resource]

ciRemainingResources :: [Resource]

ciRunningTasksCount :: Maybe Int

ciStatus :: Maybe Text

ciAgentConnected :: Lens' ContainerInstance (Maybe Bool)

This parameter returns true if the agent is actually connected to Amazon ECS. Registered instances with an agent that may be unhealthy or stopped will return false, and instances without a connected agent cannot accept placement request.

ciContainerInstanceArn :: Lens' ContainerInstance (Maybe Text)

The Amazon Resource Name (ARN) of the container instance. The ARN contains the 'arn:aws:ecs' namespace, followed by the region of the container instance, the AWS account ID of the container instance owner, the 'container-instance' namespace, and then the container instance UUID. For example, arn:aws:ecs:region:aws_account_id:container-instance/container_instance_UUID.

ciEc2InstanceId :: Lens' ContainerInstance (Maybe Text)

The Amazon EC2 instance ID of the container instance.

ciPendingTasksCount :: Lens' ContainerInstance (Maybe Int)

The number of tasks on the container instance that are in the PENDING status.

ciRegisteredResources :: Lens' ContainerInstance [Resource]

The registered resources on the container instance that are in use by current tasks.

ciRemainingResources :: Lens' ContainerInstance [Resource]

The remaining resources of the container instance that are available for new tasks.

ciRunningTasksCount :: Lens' ContainerInstance (Maybe Int)

The number of tasks on the container instance that are in the RUNNING status.

ciStatus :: Lens' ContainerInstance (Maybe Text)

The status of the container instance. The valid values are ACTIVE or INACTIVE. ACTIVE indicates that the container instance can accept tasks.

ServiceEvent

data ServiceEvent

serviceEvent :: ServiceEvent

ServiceEvent constructor.

The fields accessible through corresponding lenses are:

seCreatedAt :: Maybe UTCTime

seId :: Maybe Text

seMessage :: Maybe Text

seCreatedAt :: Lens' ServiceEvent (Maybe UTCTime)

The Unix time in seconds and milliseconds when the event was triggered.

seId :: Lens' ServiceEvent (Maybe Text)

The ID string of the event.

seMessage :: Lens' ServiceEvent (Maybe Text)

The event message.

Deployment

data Deployment

deployment :: Deployment

Deployment constructor.

The fields accessible through corresponding lenses are:

dCreatedAt :: Maybe UTCTime

dDesiredCount :: Maybe Int

dId :: Maybe Text

dPendingCount :: Maybe Int

dRunningCount :: Maybe Int

dStatus :: Maybe Text

dTaskDefinition :: Maybe Text

dUpdatedAt :: Maybe UTCTime

dCreatedAt :: Lens' Deployment (Maybe UTCTime)

The Unix time in seconds and milliseconds when the service was created.

dDesiredCount :: Lens' Deployment (Maybe Int)

The most recent desired count of tasks that was specified for the service to deploy and/or maintain.

dId :: Lens' Deployment (Maybe Text)

The ID of the deployment.

dPendingCount :: Lens' Deployment (Maybe Int)

The number of tasks in the deployment that are in the PENDING status.

dRunningCount :: Lens' Deployment (Maybe Int)

The number of tasks in the deployment that are in the RUNNING status.

dStatus :: Lens' Deployment (Maybe Text)

The status of the deployment. Valid values are PRIMARY (for the most recent deployment), ACTIVE (for previous deployments that still have tasks running, but are being replaced with the PRIMARY deployment), and INACTIVE (for deployments that have been completely replaced).

dTaskDefinition :: Lens' Deployment (Maybe Text)

The most recent task definition that was specified for the service to use.

dUpdatedAt :: Lens' Deployment (Maybe UTCTime)

The Unix time in seconds and milliseconds when the service was last updated.

MountPoint

data MountPoint

mountPoint :: MountPoint

MountPoint constructor.

The fields accessible through corresponding lenses are:

mpContainerPath :: Maybe Text

mpReadOnly :: Maybe Bool

mpSourceVolume :: Maybe Text

mpContainerPath :: Lens' MountPoint (Maybe Text)

The path on the container to mount the host volume at.

mpReadOnly :: Lens' MountPoint (Maybe Bool)

If this value is true, the container has read-only access to the volume. If this value is false, then the container can write to the volume. The default value is false.

mpSourceVolume :: Lens' MountPoint (Maybe Text)

The name of the volume to mount.