Network.AWS.ECS.UpdateContainerAgent

Description

Updates the Amazon ECS container agent on a specified container instance. Updating the Amazon ECS container agent does not interrupt running tasks or services on the container instance. The process for updating the agent differs depending on whether your container instance was launched with the Amazon ECS-optimized AMI or another operating system.

UpdateContainerAgent requires the Amazon ECS-optimized AMI or Amazon Linux with the ecs-init service installed and running. For help updating the Amazon ECS container agent on other operating systems, see Manually Updating the Amazon ECS Container Agent in the Amazon EC2 Container Service Developer Guide .

Creating a Request

updateContainerAgent

Arguments

:: Text
ucaContainerInstance

-> UpdateContainerAgent

Creates a value of UpdateContainerAgent with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ucaCluster - The short name or full Amazon Resource Name (ARN) of the cluster that your container instance is running on. If you do not specify a cluster, the default cluster is assumed.

ucaContainerInstance - The container instance ID or full Amazon Resource Name (ARN) entries for the container instance on which you would like to update the Amazon ECS container agent.

data UpdateContainerAgent

See: updateContainerAgent smart constructor.

Request Lenses

ucaCluster :: Lens' UpdateContainerAgent (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster that your container instance is running on. If you do not specify a cluster, the default cluster is assumed.

ucaContainerInstance :: Lens' UpdateContainerAgent Text

The container instance ID or full Amazon Resource Name (ARN) entries for the container instance on which you would like to update the Amazon ECS container agent.

Destructuring the Response

updateContainerAgentResponse

Arguments

:: Int
ucarsResponseStatus

-> UpdateContainerAgentResponse

Creates a value of UpdateContainerAgentResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ucarsContainerInstance - The container instance for which the container agent was updated.

ucarsResponseStatus - -- | The response status code.

data UpdateContainerAgentResponse

See: updateContainerAgentResponse smart constructor.

Response Lenses

ucarsContainerInstance :: Lens' UpdateContainerAgentResponse (Maybe ContainerInstance)

The container instance for which the container agent was updated.

ucarsResponseStatus :: Lens' UpdateContainerAgentResponse Int

- | The response status code.