Network.AWS.ECS.UpdateContainerInstancesState

Description

Modifies the status of an Amazon ECS container instance.

You can change the status of a container instance to DRAINING to manually remove an instance from a cluster, for example to perform system updates, update the Docker daemon, or scale down the cluster size.

When you set a container instance to DRAINING , Amazon ECS prevents new tasks from being scheduled for placement on the container instance and replacement service tasks are started on other container instances in the cluster if the resources are available. Service tasks on the container instance that are in the PENDING state are stopped immediately.

Service tasks on the container instance that are in the RUNNING state are stopped and replaced according the service's deployment configuration parameters, minimumHealthyPercent and maximumPercent . Note that you can change the deployment configuration of your service using UpdateService .

If minimumHealthyPercent is below 100%, the scheduler can ignore desiredCount temporarily during task replacement. For example, desiredCount is four tasks, a minimum of 50% allows the scheduler to stop two existing tasks before starting two new tasks. If the minimum is 100%, the service scheduler can't remove existing tasks until the replacement tasks are considered healthy. Tasks for services that do not use a load balancer are considered healthy if they are in the RUNNING state. Tasks for services that use a load balancer are considered healthy if they are in the RUNNING state and the container instance they are hosted on is reported as healthy by the load balancer.

The maximumPercent parameter represents an upper limit on the number of running tasks during task replacement, which enables you to define the replacement batch size. For example, if desiredCount of four tasks, a maximum of 200% starts four new tasks before stopping the four tasks to be drained (provided that the cluster resources required to do this are available). If the maximum is 100%, then replacement tasks can't start until the draining tasks have stopped.

Any PENDING or RUNNING tasks that do not belong to a service are not affected; you must wait for them to finish or stop them manually.

A container instance has completed draining when it has no more RUNNING tasks. You can verify this using ListTasks .

When you set a container instance to ACTIVE , the Amazon ECS scheduler can begin scheduling tasks on the instance again.

Creating a Request

updateContainerInstancesState

Arguments

:: ContainerInstanceStatus
ucisStatus

-> UpdateContainerInstancesState

Creates a value of UpdateContainerInstancesState with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ucisCluster - The short name or full Amazon Resource Name (ARN) of the cluster that hosts the container instance to update. If you do not specify a cluster, the default cluster is assumed.

ucisContainerInstances - A list of container instance IDs or full Amazon Resource Name (ARN) entries.

ucisStatus - The container instance state with which to update the container instance.

data UpdateContainerInstancesState

See: updateContainerInstancesState smart constructor.

Request Lenses

ucisCluster :: Lens' UpdateContainerInstancesState (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the container instance to update. If you do not specify a cluster, the default cluster is assumed.

ucisContainerInstances :: Lens' UpdateContainerInstancesState [Text]

A list of container instance IDs or full Amazon Resource Name (ARN) entries.

ucisStatus :: Lens' UpdateContainerInstancesState ContainerInstanceStatus

The container instance state with which to update the container instance.

Destructuring the Response

updateContainerInstancesStateResponse

Arguments

:: Int
ucisrsResponseStatus

-> UpdateContainerInstancesStateResponse

Creates a value of UpdateContainerInstancesStateResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ucisrsFailures - Any failures associated with the call.

ucisrsContainerInstances - The list of container instances.

ucisrsResponseStatus - -- | The response status code.

data UpdateContainerInstancesStateResponse

See: updateContainerInstancesStateResponse smart constructor.

Response Lenses

ucisrsFailures :: Lens' UpdateContainerInstancesStateResponse [Failure]

Any failures associated with the call.

ucisrsContainerInstances :: Lens' UpdateContainerInstancesStateResponse [ContainerInstance]

The list of container instances.

ucisrsResponseStatus :: Lens' UpdateContainerInstancesStateResponse Int

- | The response status code.