Network.AWS.ECS.UpdateService

Description

Modify the desired count or task definition used in a service.

You can add to or subtract from the number of instantiations of a task definition in a service by specifying the cluster that the service is running in and a new desiredCount parameter.

You can use UpdateService to modify your task definition and deploy a new version of your service, one task at a time. If you modify the task definition with UpdateService, Amazon ECS spawns a task with the new version of the task definition and then stops an old task after the new version is running. Because UpdateService starts a new version of the task before stopping an old version, your cluster must have capacity to support one more instantiation of the task when UpdateService is run. If your cluster cannot support another instantiation of the task used in your service, you can reduce the desired count of your service by one before modifying the task definition.

http://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_UpdateService.html

Request

data UpdateService

Request constructor

updateService

Arguments

:: Text
usService

-> UpdateService

UpdateService constructor.

The fields accessible through corresponding lenses are:

usCluster :: Maybe Text

usDesiredCount :: Maybe Int

usService :: Text

usTaskDefinition :: Maybe Text

Request lenses

usCluster :: Lens' UpdateService (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster that your service is running on. If you do not specify a cluster, the default cluster is assumed.

usDesiredCount :: Lens' UpdateService (Maybe Int)

The number of instantiations of the task that you would like to place and keep running in your service.

usService :: Lens' UpdateService Text

The name of the service that you want to update.

usTaskDefinition :: Lens' UpdateService (Maybe Text)

The family and revision ('family:revision') or full Amazon Resource Name (ARN) of the task definition that you want to run in your service. If you modify the task definition with UpdateService, Amazon ECS spawns a task with the new version of the task definition and then stops an old task after the new version is running.

Response

data UpdateServiceResponse

Response constructor

updateServiceResponse :: UpdateServiceResponse

UpdateServiceResponse constructor.

The fields accessible through corresponding lenses are:

usrService :: Maybe ContainerService

Response lenses

usrService :: Lens' UpdateServiceResponse (Maybe ContainerService)

The full description of your service following the update call.