Network.AWS.ECS.DeregisterContainerInstance

Description

Deregisters an Amazon ECS container instance from the specified cluster. This instance is no longer available to run tasks.

If you intend to use the container instance for some other purpose after deregistration, you should stop all of the tasks running on the container instance before deregistration to avoid any orphaned tasks from consuming resources.

Deregistering a container instance removes the instance from a cluster, but it does not terminate the EC2 instance; if you are finished using the instance, be sure to terminate it in the Amazon EC2 console to stop billing.

Creating a Request

deregisterContainerInstance

Arguments

:: Text
derContainerInstance

-> DeregisterContainerInstance

Creates a value of DeregisterContainerInstance with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

derCluster - The short name or full Amazon Resource Name (ARN) of the cluster that hosts the container instance to deregister. If you do not specify a cluster, the default cluster is assumed.

derForce - Forces the deregistration of the container instance. If you have tasks running on the container instance when you deregister it with the force option, these tasks remain running until you terminate the instance or the tasks stop through some other means, but they are orphaned (no longer monitored or accounted for by Amazon ECS). If an orphaned task on your container instance is part of an Amazon ECS service, then the service scheduler starts another copy of that task, on a different container instance if possible. Any containers in orphaned service tasks that are registered with a Classic load balancer or an Application load balancer target group are deregistered, and they will begin connection draining according to the settings on the load balancer or target group.

derContainerInstance - The container instance ID or full Amazon Resource Name (ARN) of the container instance to deregister. The ARN contains the arn:aws:ecs namespace, followed by the region of the container instance, the AWS account ID of the container instance owner, the container-instance namespace, and then the container instance ID. For example, arn:aws:ecs:region :aws_account_id :container-instance/container_instance_ID .

data DeregisterContainerInstance

See: deregisterContainerInstance smart constructor.

Request Lenses

derCluster :: Lens' DeregisterContainerInstance (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the container instance to deregister. If you do not specify a cluster, the default cluster is assumed.

derForce :: Lens' DeregisterContainerInstance (Maybe Bool)

Forces the deregistration of the container instance. If you have tasks running on the container instance when you deregister it with the force option, these tasks remain running until you terminate the instance or the tasks stop through some other means, but they are orphaned (no longer monitored or accounted for by Amazon ECS). If an orphaned task on your container instance is part of an Amazon ECS service, then the service scheduler starts another copy of that task, on a different container instance if possible. Any containers in orphaned service tasks that are registered with a Classic load balancer or an Application load balancer target group are deregistered, and they will begin connection draining according to the settings on the load balancer or target group.

derContainerInstance :: Lens' DeregisterContainerInstance Text

The container instance ID or full Amazon Resource Name (ARN) of the container instance to deregister. The ARN contains the arn:aws:ecs namespace, followed by the region of the container instance, the AWS account ID of the container instance owner, the container-instance namespace, and then the container instance ID. For example, arn:aws:ecs:region :aws_account_id :container-instance/container_instance_ID .

Destructuring the Response

deregisterContainerInstanceResponse

Arguments

:: Int
dcirsResponseStatus

-> DeregisterContainerInstanceResponse

Creates a value of DeregisterContainerInstanceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcirsContainerInstance - The container instance that was deregistered.

dcirsResponseStatus - -- | The response status code.

data DeregisterContainerInstanceResponse

See: deregisterContainerInstanceResponse smart constructor.

Response Lenses

dcirsContainerInstance :: Lens' DeregisterContainerInstanceResponse (Maybe ContainerInstance)

The container instance that was deregistered.

dcirsResponseStatus :: Lens' DeregisterContainerInstanceResponse Int

- | The response status code.