Network.AWS.ECS.DeregisterTaskDefinition

Description

Deregisters the specified task definition by family and revision. Upon deregistration, the task definition is marked as INACTIVE . Existing tasks and services that reference an INACTIVE task definition continue to run without disruption. Existing services that reference an INACTIVE task definition can still scale up or down by modifying the service's desired count.

You cannot use an INACTIVE task definition to run new tasks or create new services, and you cannot update an existing service to reference an INACTIVE task definition (although there may be up to a 10 minute window following deregistration where these restrictions have not yet taken effect).

Creating a Request

deregisterTaskDefinition

Arguments

:: Text
derTaskDefinition

-> DeregisterTaskDefinition

Creates a value of DeregisterTaskDefinition with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

derTaskDefinition - The family and revision (family:revision ) or full Amazon Resource Name (ARN) of the task definition to deregister. You must specify a revision .

data DeregisterTaskDefinition

See: deregisterTaskDefinition smart constructor.

Request Lenses

derTaskDefinition :: Lens' DeregisterTaskDefinition Text

The family and revision (family:revision ) or full Amazon Resource Name (ARN) of the task definition to deregister. You must specify a revision .

Destructuring the Response

deregisterTaskDefinitionResponse

Arguments

:: Int
dtdrsResponseStatus

-> DeregisterTaskDefinitionResponse

Creates a value of DeregisterTaskDefinitionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtdrsTaskDefinition - The full description of the deregistered task.

dtdrsResponseStatus - -- | The response status code.

data DeregisterTaskDefinitionResponse

See: deregisterTaskDefinitionResponse smart constructor.

Response Lenses

dtdrsTaskDefinition :: Lens' DeregisterTaskDefinitionResponse (Maybe TaskDefinition)

The full description of the deregistered task.

dtdrsResponseStatus :: Lens' DeregisterTaskDefinitionResponse Int

- | The response status code.