Network.AWS.ECS.DescribeTasks

Description

Describes a specified task or tasks.

Creating a Request

describeTasks :: DescribeTasks

Creates a value of DescribeTasks with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtCluster - The short name or full Amazon Resource Name (ARN) of the cluster that hosts the task to describe. If you do not specify a cluster, the default cluster is assumed.

dtTasks - A list of up to 100 task IDs or full Amazon Resource Name (ARN) entries.

data DescribeTasks

See: describeTasks smart constructor.

Request Lenses

dtCluster :: Lens' DescribeTasks (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster that hosts the task to describe. If you do not specify a cluster, the default cluster is assumed.

dtTasks :: Lens' DescribeTasks [Text]

A list of up to 100 task IDs or full Amazon Resource Name (ARN) entries.

Destructuring the Response

describeTasksResponse

Arguments

:: Int
dtrsResponseStatus

-> DescribeTasksResponse

Creates a value of DescribeTasksResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtrsFailures - Any failures associated with the call.

dtrsTasks - The list of tasks.

dtrsResponseStatus - -- | The response status code.

data DescribeTasksResponse

See: describeTasksResponse smart constructor.

Response Lenses

dtrsFailures :: Lens' DescribeTasksResponse [Failure]

Any failures associated with the call.

dtrsTasks :: Lens' DescribeTasksResponse [Task]

The list of tasks.

dtrsResponseStatus :: Lens' DescribeTasksResponse Int

- | The response status code.