Network.AWS.ECS.ListClusters

Description

Returns a list of existing clusters.

This operation returns paginated results.

Creating a Request

listClusters :: ListClusters

Creates a value of ListClusters with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lcNextToken - The nextToken value returned from a previous paginated ListClusters request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

lcMaxResults - The maximum number of cluster results returned by ListClusters in paginated output. When this parameter is used, ListClusters only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListClusters request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListClusters returns up to 100 results and a nextToken value if applicable.

data ListClusters

See: listClusters smart constructor.

Request Lenses

lcNextToken :: Lens' ListClusters (Maybe Text)

The nextToken value returned from a previous paginated ListClusters request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

lcMaxResults :: Lens' ListClusters (Maybe Int)

The maximum number of cluster results returned by ListClusters in paginated output. When this parameter is used, ListClusters only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListClusters request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListClusters returns up to 100 results and a nextToken value if applicable.

Destructuring the Response

listClustersResponse

Arguments

:: Int
lcrsResponseStatus

-> ListClustersResponse

Creates a value of ListClustersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lcrsClusterARNs - The list of full Amazon Resource Name (ARN) entries for each cluster associated with your account.

lcrsNextToken - The nextToken value to include in a future ListClusters request. When the results of a ListClusters request exceed maxResults , this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

lcrsResponseStatus - -- | The response status code.

data ListClustersResponse

See: listClustersResponse smart constructor.

Response Lenses

lcrsClusterARNs :: Lens' ListClustersResponse [Text]

The list of full Amazon Resource Name (ARN) entries for each cluster associated with your account.

lcrsNextToken :: Lens' ListClustersResponse (Maybe Text)

The nextToken value to include in a future ListClusters request. When the results of a ListClusters request exceed maxResults , this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

lcrsResponseStatus :: Lens' ListClustersResponse Int

- | The response status code.