Network.AWS.ECS.ListTaskDefinitionFamilies

Description

Returns a list of task definition families that are registered to your account (which may include task definition families that no longer have any ACTIVE task definition revisions).

You can filter out task definition families that do not contain any ACTIVE task definition revisions by setting the status parameter to ACTIVE . You can also filter the results with the familyPrefix parameter.

This operation returns paginated results.

Creating a Request

listTaskDefinitionFamilies :: ListTaskDefinitionFamilies

Creates a value of ListTaskDefinitionFamilies with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltdfStatus - The task definition family status with which to filter the ListTaskDefinitionFamilies results. By default, both ACTIVE and INACTIVE task definition families are listed. If this parameter is set to ACTIVE , only task definition families that have an ACTIVE task definition revision are returned. If this parameter is set to INACTIVE , only task definition families that do not have any ACTIVE task definition revisions are returned. If you paginate the resulting output, be sure to keep the status value constant in each subsequent request.

ltdfFamilyPrefix - The familyPrefix is a string that is used to filter the results of ListTaskDefinitionFamilies . If you specify a familyPrefix , only task definition family names that begin with the familyPrefix string are returned.

ltdfNextToken - The nextToken value returned from a previous paginated ListTaskDefinitionFamilies request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

ltdfMaxResults - The maximum number of task definition family results returned by ListTaskDefinitionFamilies in paginated output. When this parameter is used, ListTaskDefinitions only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListTaskDefinitionFamilies request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListTaskDefinitionFamilies returns up to 100 results and a nextToken value if applicable.

data ListTaskDefinitionFamilies

See: listTaskDefinitionFamilies smart constructor.

Request Lenses

ltdfStatus :: Lens' ListTaskDefinitionFamilies (Maybe TaskDefinitionFamilyStatus)

The task definition family status with which to filter the ListTaskDefinitionFamilies results. By default, both ACTIVE and INACTIVE task definition families are listed. If this parameter is set to ACTIVE , only task definition families that have an ACTIVE task definition revision are returned. If this parameter is set to INACTIVE , only task definition families that do not have any ACTIVE task definition revisions are returned. If you paginate the resulting output, be sure to keep the status value constant in each subsequent request.

ltdfFamilyPrefix :: Lens' ListTaskDefinitionFamilies (Maybe Text)

The familyPrefix is a string that is used to filter the results of ListTaskDefinitionFamilies . If you specify a familyPrefix , only task definition family names that begin with the familyPrefix string are returned.

ltdfNextToken :: Lens' ListTaskDefinitionFamilies (Maybe Text)

The nextToken value returned from a previous paginated ListTaskDefinitionFamilies request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

ltdfMaxResults :: Lens' ListTaskDefinitionFamilies (Maybe Int)

The maximum number of task definition family results returned by ListTaskDefinitionFamilies in paginated output. When this parameter is used, ListTaskDefinitions only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListTaskDefinitionFamilies request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListTaskDefinitionFamilies returns up to 100 results and a nextToken value if applicable.

Destructuring the Response

listTaskDefinitionFamiliesResponse

Arguments

:: Int
ltdfrsResponseStatus

-> ListTaskDefinitionFamiliesResponse

Creates a value of ListTaskDefinitionFamiliesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltdfrsFamilies - The list of task definition family names that match the ListTaskDefinitionFamilies request.

ltdfrsNextToken - The nextToken value to include in a future ListTaskDefinitionFamilies request. When the results of a ListTaskDefinitionFamilies request exceed maxResults , this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

ltdfrsResponseStatus - -- | The response status code.

data ListTaskDefinitionFamiliesResponse

See: listTaskDefinitionFamiliesResponse smart constructor.

Response Lenses

ltdfrsFamilies :: Lens' ListTaskDefinitionFamiliesResponse [Text]

The list of task definition family names that match the ListTaskDefinitionFamilies request.

ltdfrsNextToken :: Lens' ListTaskDefinitionFamiliesResponse (Maybe Text)

The nextToken value to include in a future ListTaskDefinitionFamilies request. When the results of a ListTaskDefinitionFamilies request exceed maxResults , this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

ltdfrsResponseStatus :: Lens' ListTaskDefinitionFamiliesResponse Int

- | The response status code.