Network.AWS.ECS.ListTaskDefinitions

Description

Returns a list of task definitions that are registered to your account. You can filter the results by family name with the familyPrefix parameter or by status with the status parameter.

This operation returns paginated results.

Creating a Request

listTaskDefinitions :: ListTaskDefinitions

Creates a value of ListTaskDefinitions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltdStatus - The task definition status with which to filter the ListTaskDefinitions results. By default, only ACTIVE task definitions are listed. By setting this parameter to INACTIVE , you can view task definitions that are INACTIVE as long as an active task or service still references them. If you paginate the resulting output, be sure to keep the status value constant in each subsequent request.

ltdFamilyPrefix - The full family name with which to filter the ListTaskDefinitions results. Specifying a familyPrefix limits the listed task definitions to task definition revisions that belong to that family.

ltdNextToken - The nextToken value returned from a previous paginated ListTaskDefinitions request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

ltdSort - The order in which to sort the results. Valid values are ASC and DESC . By default (ASC ), task definitions are listed lexicographically by family name and in ascending numerical order by revision so that the newest task definitions in a family are listed last. Setting this parameter to DESC reverses the sort order on family name and revision so that the newest task definitions in a family are listed first.

ltdMaxResults - The maximum number of task definition results returned by ListTaskDefinitions in paginated output. When this parameter is used, ListTaskDefinitions only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListTaskDefinitions request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListTaskDefinitions returns up to 100 results and a nextToken value if applicable.

data ListTaskDefinitions

See: listTaskDefinitions smart constructor.

Request Lenses

ltdStatus :: Lens' ListTaskDefinitions (Maybe TaskDefinitionStatus)

The task definition status with which to filter the ListTaskDefinitions results. By default, only ACTIVE task definitions are listed. By setting this parameter to INACTIVE , you can view task definitions that are INACTIVE as long as an active task or service still references them. If you paginate the resulting output, be sure to keep the status value constant in each subsequent request.

ltdFamilyPrefix :: Lens' ListTaskDefinitions (Maybe Text)

The full family name with which to filter the ListTaskDefinitions results. Specifying a familyPrefix limits the listed task definitions to task definition revisions that belong to that family.

ltdNextToken :: Lens' ListTaskDefinitions (Maybe Text)

The nextToken value returned from a previous paginated ListTaskDefinitions request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

ltdSort :: Lens' ListTaskDefinitions (Maybe SortOrder)

The order in which to sort the results. Valid values are ASC and DESC . By default (ASC ), task definitions are listed lexicographically by family name and in ascending numerical order by revision so that the newest task definitions in a family are listed last. Setting this parameter to DESC reverses the sort order on family name and revision so that the newest task definitions in a family are listed first.

ltdMaxResults :: Lens' ListTaskDefinitions (Maybe Int)

The maximum number of task definition results returned by ListTaskDefinitions in paginated output. When this parameter is used, ListTaskDefinitions only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListTaskDefinitions request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListTaskDefinitions returns up to 100 results and a nextToken value if applicable.

Destructuring the Response

listTaskDefinitionsResponse

Arguments

:: Int
ltdrsResponseStatus

-> ListTaskDefinitionsResponse

Creates a value of ListTaskDefinitionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltdrsTaskDefinitionARNs - The list of task definition Amazon Resource Name (ARN) entries for the ListTaskDefinitions request.

ltdrsNextToken - The nextToken value to include in a future ListTaskDefinitions request. When the results of a ListTaskDefinitions request exceed maxResults , this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

ltdrsResponseStatus - -- | The response status code.

data ListTaskDefinitionsResponse

See: listTaskDefinitionsResponse smart constructor.

Response Lenses

ltdrsTaskDefinitionARNs :: Lens' ListTaskDefinitionsResponse [Text]

The list of task definition Amazon Resource Name (ARN) entries for the ListTaskDefinitions request.

ltdrsNextToken :: Lens' ListTaskDefinitionsResponse (Maybe Text)

The nextToken value to include in a future ListTaskDefinitions request. When the results of a ListTaskDefinitions request exceed maxResults , this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

ltdrsResponseStatus :: Lens' ListTaskDefinitionsResponse Int

- | The response status code.