Network.AWS.ECS.RegisterContainerInstance

Description

Registers an EC2 instance into the specified cluster. This instance becomes available to place containers on.

Creating a Request

registerContainerInstance :: RegisterContainerInstance

Creates a value of RegisterContainerInstance with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rciInstanceIdentityDocumentSignature - The instance identity document signature for the EC2 instance to register. This signature can be found by running the following command from the instance: curl http://169.254.169.254/latest/dynamic/instance-identity/signature/

rciCluster - The short name or full Amazon Resource Name (ARN) of the cluster with which to register your container instance. If you do not specify a cluster, the default cluster is assumed.

rciInstanceIdentityDocument - The instance identity document for the EC2 instance to register. This document can be found by running the following command from the instance: curl http://169.254.169.254/latest/dynamic/instance-identity/document/

rciContainerInstanceARN - The Amazon Resource Name (ARN) of the container instance (if it was previously registered).

rciVersionInfo - The version information for the Amazon ECS container agent and Docker daemon running on the container instance.

rciAttributes - The container instance attributes that this container instance supports.

rciTotalResources - The resources available on the instance.

data RegisterContainerInstance

See: registerContainerInstance smart constructor.

Request Lenses

rciInstanceIdentityDocumentSignature :: Lens' RegisterContainerInstance (Maybe Text)

The instance identity document signature for the EC2 instance to register. This signature can be found by running the following command from the instance: curl http://169.254.169.254/latest/dynamic/instance-identity/signature/

rciCluster :: Lens' RegisterContainerInstance (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster with which to register your container instance. If you do not specify a cluster, the default cluster is assumed.

rciInstanceIdentityDocument :: Lens' RegisterContainerInstance (Maybe Text)

The instance identity document for the EC2 instance to register. This document can be found by running the following command from the instance: curl http://169.254.169.254/latest/dynamic/instance-identity/document/

rciContainerInstanceARN :: Lens' RegisterContainerInstance (Maybe Text)

The Amazon Resource Name (ARN) of the container instance (if it was previously registered).

rciVersionInfo :: Lens' RegisterContainerInstance (Maybe VersionInfo)

The version information for the Amazon ECS container agent and Docker daemon running on the container instance.

rciAttributes :: Lens' RegisterContainerInstance [Attribute]

The container instance attributes that this container instance supports.

rciTotalResources :: Lens' RegisterContainerInstance [Resource]

The resources available on the instance.

Destructuring the Response

registerContainerInstanceResponse

Arguments

:: Int
rcirsResponseStatus

-> RegisterContainerInstanceResponse

Creates a value of RegisterContainerInstanceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rcirsContainerInstance - The container instance that was registered.

rcirsResponseStatus - -- | The response status code.

data RegisterContainerInstanceResponse

See: registerContainerInstanceResponse smart constructor.

Response Lenses

rcirsContainerInstance :: Lens' RegisterContainerInstanceResponse (Maybe ContainerInstance)

The container instance that was registered.

rcirsResponseStatus :: Lens' RegisterContainerInstanceResponse Int

- | The response status code.