Network.AWS.ECS.RegisterTaskDefinition

Description

Registers a new task definition from the supplied family and containerDefinitions . Optionally, you can add data volumes to your containers with the volumes parameter. For more information about task definition parameters and defaults, see Amazon ECS Task Definitions in the Amazon EC2 Container Service Developer Guide .

You can specify an IAM role for your task with the taskRoleArn parameter. When you specify an IAM role for a task, its containers can then use the latest versions of the AWS CLI or SDKs to make API requests to the AWS services that are specified in the IAM policy associated with the role. For more information, see IAM Roles for Tasks in the Amazon EC2 Container Service Developer Guide .

You can specify a Docker networking mode for the containers in your task definition with the networkMode parameter. The available network modes correspond to those described in Network settings in the Docker run reference.

Creating a Request

registerTaskDefinition

Arguments

:: Text
rtdFamily

-> RegisterTaskDefinition

Creates a value of RegisterTaskDefinition with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rtdTaskRoleARN - The short name or full Amazon Resource Name (ARN) of the IAM role that containers in this task can assume. All containers in this task are granted the permissions that are specified in this role. For more information, see IAM Roles for Tasks in the Amazon EC2 Container Service Developer Guide .

rtdPlacementConstraints - An array of placement constraint objects to use for the task. You can specify a maximum of 10 constraints per task (this limit includes constraints in the task definition and those specified at run time).

rtdNetworkMode - The Docker networking mode to use for the containers in the task. The valid values are none , bridge , and host . The default Docker network mode is bridge . If the network mode is set to none , you cannot specify port mappings in your container definitions, and the task's containers do not have external connectivity. The host network mode offers the highest networking performance for containers because they use the host network stack instead of the virtualized network stack provided by the bridge mode; however, exposed container ports are mapped directly to the corresponding host port, so you cannot take advantage of dynamic host port mappings or run multiple instantiations of the same task on a single container instance if port mappings are used. For more information, see Network settings in the Docker run reference .

rtdVolumes - A list of volume definitions in JSON format that containers in your task may use.

rtdFamily - You must specify a family for a task definition, which allows you to track multiple versions of the same task definition. The family is used as a name for your task definition. Up to 255 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed.

rtdContainerDefinitions - A list of container definitions in JSON format that describe the different containers that make up your task.

data RegisterTaskDefinition

See: registerTaskDefinition smart constructor.

Request Lenses

rtdTaskRoleARN :: Lens' RegisterTaskDefinition (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the IAM role that containers in this task can assume. All containers in this task are granted the permissions that are specified in this role. For more information, see IAM Roles for Tasks in the Amazon EC2 Container Service Developer Guide .

rtdPlacementConstraints :: Lens' RegisterTaskDefinition [TaskDefinitionPlacementConstraint]

An array of placement constraint objects to use for the task. You can specify a maximum of 10 constraints per task (this limit includes constraints in the task definition and those specified at run time).

rtdNetworkMode :: Lens' RegisterTaskDefinition (Maybe NetworkMode)

The Docker networking mode to use for the containers in the task. The valid values are none , bridge , and host . The default Docker network mode is bridge . If the network mode is set to none , you cannot specify port mappings in your container definitions, and the task's containers do not have external connectivity. The host network mode offers the highest networking performance for containers because they use the host network stack instead of the virtualized network stack provided by the bridge mode; however, exposed container ports are mapped directly to the corresponding host port, so you cannot take advantage of dynamic host port mappings or run multiple instantiations of the same task on a single container instance if port mappings are used. For more information, see Network settings in the Docker run reference .

rtdVolumes :: Lens' RegisterTaskDefinition [Volume]

A list of volume definitions in JSON format that containers in your task may use.

rtdFamily :: Lens' RegisterTaskDefinition Text

You must specify a family for a task definition, which allows you to track multiple versions of the same task definition. The family is used as a name for your task definition. Up to 255 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed.

rtdContainerDefinitions :: Lens' RegisterTaskDefinition [ContainerDefinition]

A list of container definitions in JSON format that describe the different containers that make up your task.

Destructuring the Response

registerTaskDefinitionResponse

Arguments

:: Int
rtdrsResponseStatus

-> RegisterTaskDefinitionResponse

Creates a value of RegisterTaskDefinitionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rtdrsTaskDefinition - The full description of the registered task definition.

rtdrsResponseStatus - -- | The response status code.

data RegisterTaskDefinitionResponse

See: registerTaskDefinitionResponse smart constructor.

Response Lenses

rtdrsTaskDefinition :: Lens' RegisterTaskDefinitionResponse (Maybe TaskDefinition)

The full description of the registered task definition.

rtdrsResponseStatus :: Lens' RegisterTaskDefinitionResponse Int

- | The response status code.