Network.AWS.ECS.RunTask

Description

Starts a new task using the specified task definition.

You can allow Amazon ECS to place tasks for you, or you can customize how Amazon ECS places tasks using placement constraints and placement strategies. For more information, see Scheduling Tasks in the Amazon EC2 Container Service Developer Guide .

Alternatively, you can use StartTask to use your own scheduler or place tasks manually on specific container instances.

Creating a Request

runTask

Arguments

:: Text
rtTaskDefinition

-> RunTask

Creates a value of RunTask with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rtOverrides - A list of container overrides in JSON format that specify the name of a container in the specified task definition and the overrides it should receive. You can override the default command for a container (that is specified in the task definition or Docker image) with a command override. You can also override existing environment variables (that are specified in the task definition or Docker image) on a container or add new environment variables to it with an environment override.

rtGroup - The name of the task group to associate with the task. The default value is the family name of the task definition (for example, family:my-family-name).

rtCluster - The short name or full Amazon Resource Name (ARN) of the cluster on which to run your task. If you do not specify a cluster, the default cluster is assumed.

rtCount - The number of instantiations of the specified task to place on your cluster. You can specify up to 10 tasks per call.

rtPlacementConstraints - An array of placement constraint objects to use for the task. You can specify up to 10 constraints per task (including constraints in the task definition and those specified at run time).

rtPlacementStrategy - The placement strategy objects to use for the task. You can specify a maximum of 5 strategy rules per task.

rtStartedBy - An optional tag specified when a task is started. For example if you automatically trigger a task to run a batch process job, you could apply a unique identifier for that job to your task with the startedBy parameter. You can then identify which tasks belong to that job by filtering the results of a ListTasks call with the startedBy value. Up to 36 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed. If a task is started by an Amazon ECS service, then the startedBy parameter contains the deployment ID of the service that starts it.

rtTaskDefinition - The family and revision (family:revision ) or full Amazon Resource Name (ARN) of the task definition to run. If a revision is not specified, the latest ACTIVE revision is used.

data RunTask

See: runTask smart constructor.

Request Lenses

rtOverrides :: Lens' RunTask (Maybe TaskOverride)

A list of container overrides in JSON format that specify the name of a container in the specified task definition and the overrides it should receive. You can override the default command for a container (that is specified in the task definition or Docker image) with a command override. You can also override existing environment variables (that are specified in the task definition or Docker image) on a container or add new environment variables to it with an environment override.

rtGroup :: Lens' RunTask (Maybe Text)

The name of the task group to associate with the task. The default value is the family name of the task definition (for example, family:my-family-name).

rtCluster :: Lens' RunTask (Maybe Text)

The short name or full Amazon Resource Name (ARN) of the cluster on which to run your task. If you do not specify a cluster, the default cluster is assumed.

rtCount :: Lens' RunTask (Maybe Int)

The number of instantiations of the specified task to place on your cluster. You can specify up to 10 tasks per call.

rtPlacementConstraints :: Lens' RunTask [PlacementConstraint]

An array of placement constraint objects to use for the task. You can specify up to 10 constraints per task (including constraints in the task definition and those specified at run time).

rtPlacementStrategy :: Lens' RunTask [PlacementStrategy]

The placement strategy objects to use for the task. You can specify a maximum of 5 strategy rules per task.

rtStartedBy :: Lens' RunTask (Maybe Text)

An optional tag specified when a task is started. For example if you automatically trigger a task to run a batch process job, you could apply a unique identifier for that job to your task with the startedBy parameter. You can then identify which tasks belong to that job by filtering the results of a ListTasks call with the startedBy value. Up to 36 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed. If a task is started by an Amazon ECS service, then the startedBy parameter contains the deployment ID of the service that starts it.

rtTaskDefinition :: Lens' RunTask Text

The family and revision (family:revision ) or full Amazon Resource Name (ARN) of the task definition to run. If a revision is not specified, the latest ACTIVE revision is used.

Destructuring the Response

runTaskResponse

Arguments

:: Int
rtrsResponseStatus

-> RunTaskResponse

Creates a value of RunTaskResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rtrsFailures - Any failures associated with the call.

rtrsTasks - A full description of the tasks that were run. Each task that was successfully placed on your cluster are described here.

rtrsResponseStatus - -- | The response status code.

data RunTaskResponse

See: runTaskResponse smart constructor.

Response Lenses

rtrsFailures :: Lens' RunTaskResponse [Failure]

Any failures associated with the call.

rtrsTasks :: Lens' RunTaskResponse [Task]

A full description of the tasks that were run. Each task that was successfully placed on your cluster are described here.

rtrsResponseStatus :: Lens' RunTaskResponse Int

- | The response status code.